refactor: simplify tree registry components
