+++ apps/www/__registry__/index.tsx
@@ -521,7 +521,7 @@ export const index: Record<string, any> = {
         type: 'registry:ui',
         target: 'components/jetbrains-ui/scroll-area.tsx',
         content:
-          "import { ScrollArea as ScrollAreaPrimitive } from 'radix-ui';\nimport * as React from 'react';\n\nimport { cn } from '@/lib/utils';\n\nconst ScrollArea = React.forwardRef<\n  React.ComponentRef<typeof ScrollAreaPrimitive.Root>,\n  React.ComponentPropsWithoutRef<typeof ScrollAreaPrimitive.Root>\n>(({ className, children, ...props }, ref) => (\n  <ScrollAreaPrimitive.Root\n    ref={ref}\n    className={cn('overflow-hidden', className)}\n    {...props}\n  >\n    {children}\n    <ScrollAreaPrimitive.Corner />\n    <ScrollBar orientation=\"vertical\" />\n  </ScrollAreaPrimitive.Root>\n));\n\nScrollArea.displayName = ScrollAreaPrimitive.Root.displayName;\n\nconst ScrollViewport = React.forwardRef<\n  React.ComponentRef<typeof ScrollAreaPrimitive.Viewport>,\n  React.ComponentPropsWithoutRef<typeof ScrollAreaPrimitive.Viewport>\n>(({ className, children, ...props }, ref) => (\n  <ScrollAreaPrimitive.Viewport\n    ref={ref}\n    className={cn('size-full rounded-[inherit]', className)}\n    {...props}\n  >\n    {children}\n  </ScrollAreaPrimitive.Viewport>\n));\n\nScrollViewport.displayName = ScrollAreaPrimitive.Viewport.displayName;\n\nconst ScrollBar = React.forwardRef<\n  React.ComponentRef<typeof ScrollAreaPrimitive.Scrollbar>,\n  React.ComponentPropsWithoutRef<typeof ScrollAreaPrimitive.Scrollbar>\n>(({ className, orientation = 'vertical', ...props }, ref) => (\n  <ScrollAreaPrimitive.Scrollbar\n    ref={ref}\n    orientation={orientation}\n    className={cn(\n      'flex select-none data-[state=hidden]:animate-fd-fade-out',\n      orientation === 'vertical' && 'h-full w-1.5',\n      orientation === 'horizontal' && 'h-1.5 flex-col',\n      className,\n    )}\n    {...props}\n  >\n    <ScrollAreaPrimitive.ScrollAreaThumb className=\"relative flex-1 rounded-full bg-fd-border\" />\n  </ScrollAreaPrimitive.Scrollbar>\n));\nScrollBar.displayName = ScrollAreaPrimitive.Scrollbar.displayName;\n\nexport { ScrollArea, ScrollBar, ScrollViewport };",
+          "import { ScrollArea as ScrollAreaPrimitive } from 'radix-ui';\nimport * as React from 'react';\n\nimport { cn } from '@/lib/utils';\n\nconst ScrollArea = React.forwardRef<\n  React.ComponentRef<typeof ScrollAreaPrimitive.Root>,\n  React.ComponentPropsWithoutRef<typeof ScrollAreaPrimitive.Root>\n>(({ className, children, ...props }, ref) => (\n  <ScrollAreaPrimitive.Root\n    ref={ref}\n    className={cn('overflow-hidden', className)}\n    {...props}\n  >\n    {children}\n    <ScrollAreaPrimitive.Corner />\n    <ScrollBar orientation=\"horizontal\" />\n    <ScrollBar orientation=\"vertical\" />\n  </ScrollAreaPrimitive.Root>\n));\n\nScrollArea.displayName = ScrollAreaPrimitive.Root.displayName;\n\nconst ScrollViewport = React.forwardRef<\n  React.ComponentRef<typeof ScrollAreaPrimitive.Viewport>,\n  React.ComponentPropsWithoutRef<typeof ScrollAreaPrimitive.Viewport>\n>(({ className, children, ...props }, ref) => (\n  <ScrollAreaPrimitive.Viewport\n    ref={ref}\n    className={cn('size-full rounded-[inherit]', className)}\n    {...props}\n  >\n    {children}\n  </ScrollAreaPrimitive.Viewport>\n));\n\nScrollViewport.displayName = ScrollAreaPrimitive.Viewport.displayName;\n\nconst ScrollBar = React.forwardRef<\n  React.ComponentRef<typeof ScrollAreaPrimitive.Scrollbar>,\n  React.ComponentPropsWithoutRef<typeof ScrollAreaPrimitive.Scrollbar>\n>(({ className, orientation = 'vertical', ...props }, ref) => (\n  <ScrollAreaPrimitive.Scrollbar\n    ref={ref}\n    orientation={orientation}\n    className={cn(\n      'flex select-none data-[state=hidden]:animate-fd-fade-out',\n      orientation === 'vertical' && 'h-full w-1.5',\n      orientation === 'horizontal' && 'h-1.5 flex-col',\n      className,\n    )}\n    {...props}\n  >\n    <ScrollAreaPrimitive.ScrollAreaThumb className=\"relative flex-1 rounded-full bg-fd-border\" />\n  </ScrollAreaPrimitive.Scrollbar>\n));\nScrollBar.displayName = ScrollAreaPrimitive.Scrollbar.displayName;\n\nexport { ScrollArea, ScrollBar, ScrollViewport };",
       },
     ],
     keywords: ['scroll-area', 'jetbrains', 'ui', 'component'],
@@ -761,8 +761,8 @@ export const index: Record<string, any> = {
     devDependencies: undefined,
     registryDependencies: [
       'utils',
-      '@jetbrains-ui/context-menu',
       '@jetbrains-ui/icons-components',
+      '@jetbrains-ui/scroll-area',
     ],
     files: [
       {
@@ -770,21 +770,35 @@ export const index: Record<string, any> = {
         type: 'registry:ui',
         target: 'components/jetbrains-ui/tree/index.tsx',
         content:
-          "'use client';\n\nimport * as React from 'react';\n\nimport { cn } from '@/lib/utils';\nimport {\n  AddIcon,\n  CheckIcon,\n  ChevronDownIcon,\n  ChevronRightIcon,\n  CloseIcon,\n  CopyIcon,\n  EditIcon,\n  FilterIcon,\n  RefreshIcon,\n  SearchIcon,\n  SettingsIcon,\n  TrashIcon,\n} from '@/components/jetbrains-ui/icons';\nimport {\n  ContextMenu,\n  ContextMenuContent,\n  ContextMenuItem,\n  ContextMenuSeparator,\n  ContextMenuSub,\n  ContextMenuSubContent,\n  ContextMenuSubTrigger,\n  ContextMenuTrigger,\n} from '@/components/jetbrains-ui/context-menu';\nimport './index.css';\n\nconst treeContextMenuIcons = {\n  add: AddIcon,\n  check: CheckIcon,\n  close: CloseIcon,\n  copy: CopyIcon,\n  delete: TrashIcon,\n  edit: EditIcon,\n  filter: FilterIcon,\n  refresh: RefreshIcon,\n  search: SearchIcon,\n  settings: SettingsIcon,\n  trash: TrashIcon,\n} as const;\n\nexport type TreeContextMenuIconName = keyof typeof treeContextMenuIcons;\n\nexport interface TreeContextMenuNode {\n  id: string;\n  value: string;\n  label: string;\n  data: Pick<TreeItemProps, 'label' | 'icon' | 'endContent' | 'disabled'>;\n  disabled: boolean;\n  expanded: boolean;\n  hasChildren: boolean;\n  level: number;\n  selected: boolean;\n}\n\nexport interface TreeContextMenuItem {\n  key: string;\n  label?: React.ReactNode;\n  text?: React.ReactNode;\n  disabled?: boolean;\n  separator?: boolean;\n  separatorAfter?: boolean;\n  iconName?: TreeContextMenuIconName;\n  variant?: 'default' | 'destructive';\n  onSelect?: (node: TreeContextMenuNode) => void;\n  children?: TreeContextMenuItem[];\n  items?: TreeContextMenuItem[];\n}\n\ntype RegisteredTreeItem = TreeContextMenuNode;\n\ntype TreeContextValue = {\n  selectedId: string | null;\n  setSelectedId: (nextId: string) => void;\n  expandedIds: Set<string>;\n  toggleExpanded: (id: string, expanded?: boolean) => void;\n  indent: number;\n  registerItem: (item: RegisteredTreeItem) => void;\n  unregisterItem: (id: string) => void;\n};\n\ninterface TreeProps\n  extends Omit<React.ComponentPropsWithoutRef<'ul'>, 'contextMenu'> {\n  selectedId?: string | null;\n  defaultSelectedId?: string | null;\n  onSelectedIdChange?: (id: string) => void;\n  expandedIds?: string[];\n  defaultExpandedIds?: string[];\n  onExpandedIdsChange?: (ids: string[]) => void;\n  indent?: number;\n  contextMenu?: (\n    node: TreeContextMenuNode,\n  ) => TreeContextMenuItem[] | null | undefined;\n}\n\ninterface TreeItemProps\n  extends Omit<React.ComponentProps<'li'>, 'children' | 'id'> {\n  value: string;\n  label: React.ReactNode;\n  icon?: React.ReactNode;\n  endContent?: React.ReactNode;\n  disabled?: boolean;\n  children?: React.ReactNode;\n}\n\nconst TreeContext = React.createContext<TreeContextValue | null>(null);\nconst TreeLevelContext = React.createContext(1);\nconst GROUP_COLLAPSE_DURATION = 180;\nconst treeRootClassName =\n  'jb-tree block w-[244px] max-w-full box-border py-2 text-[13px] leading-4 font-medium';\nconst treeListClassName = 'm-0 list-none p-0';\nconst treeItemClassName = 'jb-tree-item m-0 list-none p-0';\nconst treeGroupClassName = 'jb-tree-group m-0 list-none p-0';\nconst treeItemButtonClassName =\n  'jb-tree-item-button relative flex h-6 w-full min-w-0 cursor-default items-center gap-0.5 pr-4 select-none outline-none transition-[color,background-color,box-shadow] duration-150 ease-in-out';\nconst treeDisclosureClassName =\n  'jb-tree-item-disclosure inline-flex size-4 shrink-0 items-center justify-center';\nconst treeDisclosurePlaceholderClassName =\n  'jb-tree-item-disclosure-placeholder inline-flex size-4 shrink-0 items-center justify-center';\nconst treeItemContentClassName =\n  'jb-tree-item-content inline-flex h-5 min-w-0 flex-1 items-center gap-1.5';\nconst treeItemIconClassName =\n  'jb-tree-item-icon inline-flex size-4 shrink-0 items-center justify-center';\nconst treeItemLabelClassName =\n  'jb-tree-item-label min-w-0 overflow-hidden text-ellipsis whitespace-nowrap';\nconst treeItemEndClassName =\n  'jb-tree-item-end ml-auto pr-0.5 text-[13px] leading-4 font-medium';\n\nfunction useTreeContext() {\n  const context = React.useContext(TreeContext);\n\n  if (!context) {\n    throw new Error('Tree components must be used within <Tree>.');\n  }\n\n  return context;\n}\n\nfunction mergeExpandedIds(\n  current: string[],\n  id: string,\n  expanded: boolean,\n): string[] {\n  const next = new Set(current);\n\n  if (expanded) {\n    next.add(id);\n  } else {\n    next.delete(id);\n  }\n\n  return Array.from(next);\n}\n\nfunction getTreeItemLabel(label: React.ReactNode, fallback: string) {\n  if (typeof label === 'string') {\n    return label;\n  }\n\n  return fallback;\n}\n\nfunction resolveContextMenuLabel(item: TreeContextMenuItem) {\n  return item.label ?? item.text ?? item.key;\n}\n\nfunction resolveContextMenuChildren(item: TreeContextMenuItem) {\n  return item.children ?? item.items ?? [];\n}\n\nfunction resolveContextMenuSeparator(item: TreeContextMenuItem) {\n  return item.separatorAfter ?? item.separator ?? false;\n}\n\nfunction TreeContextMenuEntries({\n  items,\n  node,\n}: {\n  items: TreeContextMenuItem[];\n  node: TreeContextMenuNode;\n}) {\n  return (\n    <>\n      {items.map((item) => {\n        const label = resolveContextMenuLabel(item);\n        const childItems = resolveContextMenuChildren(item);\n        const IconComponent = item.iconName\n          ? treeContextMenuIcons[item.iconName]\n          : null;\n        const entry = childItems.length ? (\n          <ContextMenuSub key={item.key}>\n            <ContextMenuSubTrigger disabled={item.disabled}>\n              {IconComponent ? (\n                <IconComponent\n                  aria-hidden=\"true\"\n                  className=\"size-4\"\n                  size={16}\n                />\n              ) : null}\n              <span className=\"truncate\">{label}</span>\n            </ContextMenuSubTrigger>\n            <ContextMenuSubContent>\n              <TreeContextMenuEntries items={childItems} node={node} />\n            </ContextMenuSubContent>\n          </ContextMenuSub>\n        ) : (\n          <ContextMenuItem\n            key={item.key}\n            disabled={item.disabled}\n            variant={item.variant}\n            onSelect={() => item.onSelect?.(node)}\n          >\n            {IconComponent ? (\n              <IconComponent aria-hidden=\"true\" className=\"size-4\" size={16} />\n            ) : null}\n            <span className=\"truncate\">{label}</span>\n          </ContextMenuItem>\n        );\n\n        if (!resolveContextMenuSeparator(item)) {\n          return entry;\n        }\n\n        return (\n          <React.Fragment key={`${item.key}-group`}>\n            {entry}\n            <ContextMenuSeparator />\n          </React.Fragment>\n        );\n      })}\n    </>\n  );\n}\n\nfunction Tree({\n  children,\n  className,\n  style,\n  selectedId,\n  defaultSelectedId = null,\n  onSelectedIdChange,\n  expandedIds,\n  defaultExpandedIds = [],\n  onExpandedIdsChange,\n  indent = 16,\n  contextMenu,\n  onContextMenuCapture,\n  ...props\n}: TreeProps) {\n  const [uncontrolledSelectedId, setUncontrolledSelectedId] = React.useState<\n    string | null\n  >(defaultSelectedId);\n  const [uncontrolledExpandedIds, setUncontrolledExpandedIds] =\n    React.useState<string[]>(defaultExpandedIds);\n  const [contextMenuNode, setContextMenuNode] =\n    React.useState<TreeContextMenuNode | null>(null);\n  const [contextMenuItems, setContextMenuItems] = React.useState<\n    TreeContextMenuItem[]\n  >([]);\n  const itemRegistryRef = React.useRef(new Map<string, RegisteredTreeItem>());\n\n  const resolvedSelectedId =\n    selectedId !== undefined ? selectedId : uncontrolledSelectedId;\n  const resolvedExpandedIds =\n    expandedIds !== undefined ? expandedIds : uncontrolledExpandedIds;\n\n  const setSelectedId = React.useCallback(\n    (nextId: string) => {\n      if (selectedId === undefined) {\n        setUncontrolledSelectedId(nextId);\n      }\n\n      onSelectedIdChange?.(nextId);\n    },\n    [onSelectedIdChange, selectedId],\n  );\n\n  const setExpandedIds = React.useCallback(\n    (nextIds: string[]) => {\n      if (expandedIds === undefined) {\n        setUncontrolledExpandedIds(nextIds);\n      }\n\n      onExpandedIdsChange?.(nextIds);\n    },\n    [expandedIds, onExpandedIdsChange],\n  );\n\n  const toggleExpanded = React.useCallback(\n    (id: string, expanded?: boolean) => {\n      const isExpanded = resolvedExpandedIds.includes(id);\n      const nextExpanded = expanded ?? !isExpanded;\n      const nextIds = mergeExpandedIds(resolvedExpandedIds, id, nextExpanded);\n\n      setExpandedIds(nextIds);\n    },\n    [resolvedExpandedIds, setExpandedIds],\n  );\n\n  const registerItem = React.useCallback((item: RegisteredTreeItem) => {\n    itemRegistryRef.current.set(item.id, item);\n  }, []);\n\n  const unregisterItem = React.useCallback((id: string) => {\n    itemRegistryRef.current.delete(id);\n  }, []);\n\n  const clearContextMenuState = React.useCallback(() => {\n    setContextMenuNode(null);\n    setContextMenuItems([]);\n  }, []);\n\n  const contextValue = React.useMemo<TreeContextValue>(\n    () => ({\n      selectedId: resolvedSelectedId,\n      setSelectedId,\n      expandedIds: new Set(resolvedExpandedIds),\n      toggleExpanded,\n      indent,\n      registerItem,\n      unregisterItem,\n    }),\n    [\n      indent,\n      registerItem,\n      resolvedExpandedIds,\n      resolvedSelectedId,\n      setSelectedId,\n      toggleExpanded,\n      unregisterItem,\n    ],\n  );\n\n  const handleTreeContextMenuCapture = React.useCallback(\n    (event: React.MouseEvent<HTMLUListElement>) => {\n      onContextMenuCapture?.(event);\n\n      if (event.defaultPrevented || !contextMenu) {\n        return;\n      }\n\n      const target = event.target;\n\n      if (!(target instanceof HTMLElement)) {\n        event.preventDefault();\n        clearContextMenuState();\n        return;\n      }\n\n      const itemButton = target.closest(\n        '[data-slot=\"jb-tree-item-button\"][data-value]',\n      );\n\n      if (!(itemButton instanceof HTMLElement)) {\n        event.preventDefault();\n        clearContextMenuState();\n        return;\n      }\n\n      const itemId = itemButton.dataset.value;\n\n      if (!itemId) {\n        event.preventDefault();\n        clearContextMenuState();\n        return;\n      }\n\n      const node = itemRegistryRef.current.get(itemId);\n\n      if (!node) {\n        event.preventDefault();\n        clearContextMenuState();\n        return;\n      }\n\n      const items = contextMenu(node) ?? [];\n\n      if (items.length === 0) {\n        event.preventDefault();\n        clearContextMenuState();\n        return;\n      }\n\n      setContextMenuNode(node);\n      setContextMenuItems(items);\n    },\n    [clearContextMenuState, contextMenu, onContextMenuCapture],\n  );\n\n  const treeList = (\n    <ul\n      data-slot=\"jb-tree\"\n      role=\"tree\"\n      className={cn(treeRootClassName, treeListClassName, className)}\n      style={style}\n      {...props}\n      onContextMenuCapture={\n        contextMenu ? handleTreeContextMenuCapture : onContextMenuCapture\n      }\n    >\n      <TreeLevelContext.Provider value={1}>\n        {children}\n      </TreeLevelContext.Provider>\n    </ul>\n  );\n\n  return (\n    <TreeContext.Provider value={contextValue}>\n      {contextMenu ? (\n        <ContextMenu\n          onOpenChange={(open) => {\n            if (!open) {\n              clearContextMenuState();\n            }\n          }}\n        >\n          <ContextMenuTrigger asChild>{treeList}</ContextMenuTrigger>\n          {contextMenuNode && contextMenuItems.length ? (\n            <ContextMenuContent>\n              <TreeContextMenuEntries\n                items={contextMenuItems}\n                node={contextMenuNode}\n              />\n            </ContextMenuContent>\n          ) : null}\n        </ContextMenu>\n      ) : (\n        treeList\n      )}\n    </TreeContext.Provider>\n  );\n}\n\nfunction TreeItem({\n  className,\n  value,\n  label,\n  icon,\n  endContent,\n  disabled = false,\n  children,\n  ...props\n}: TreeItemProps) {\n  const {\n    selectedId,\n    setSelectedId,\n    expandedIds,\n    toggleExpanded,\n    indent,\n    registerItem,\n    unregisterItem,\n  } = useTreeContext();\n  const level = React.useContext(TreeLevelContext);\n\n  const hasChildren = React.Children.count(children) > 0;\n  const expanded = hasChildren ? expandedIds.has(value) : false;\n  const selected = selectedId === value;\n  const textLabel = getTreeItemLabel(label, value);\n  const [keepGroupMounted, setKeepGroupMounted] = React.useState(expanded);\n  const [groupExpanded, setGroupExpanded] = React.useState(expanded);\n  const [isCollapsing, setIsCollapsing] = React.useState(false);\n  const hasMountedRef = React.useRef(false);\n  const collapseTimerRef = React.useRef<number | null>(null);\n  const expandRafRef = React.useRef<number | null>(null);\n\n  const rowPaddingLeft = 16 + (level - 1) * indent + Math.max(0, level - 2) * 2;\n\n  const clearGroupAnimationHandles = React.useCallback(() => {\n    if (collapseTimerRef.current !== null) {\n      window.clearTimeout(collapseTimerRef.current);\n      collapseTimerRef.current = null;\n    }\n\n    if (expandRafRef.current !== null) {\n      window.cancelAnimationFrame(expandRafRef.current);\n      expandRafRef.current = null;\n    }\n  }, []);\n\n  React.useEffect(() => {\n    return () => {\n      clearGroupAnimationHandles();\n    };\n  }, [clearGroupAnimationHandles]);\n\n  React.useEffect(() => {\n    clearGroupAnimationHandles();\n\n    if (!hasChildren) {\n      setKeepGroupMounted(false);\n      setGroupExpanded(false);\n      setIsCollapsing(false);\n      return;\n    }\n\n    if (!hasMountedRef.current) {\n      hasMountedRef.current = true;\n      setKeepGroupMounted(expanded);\n      setGroupExpanded(expanded);\n      setIsCollapsing(false);\n      return;\n    }\n\n    if (expanded) {\n      setKeepGroupMounted(true);\n      setIsCollapsing(false);\n\n      // Mount collapsed first, then expand on next frame so transition is visible.\n      setGroupExpanded(false);\n      expandRafRef.current = window.requestAnimationFrame(() => {\n        expandRafRef.current = window.requestAnimationFrame(() => {\n          setGroupExpanded(true);\n          expandRafRef.current = null;\n        });\n      });\n\n      return;\n    }\n\n    if (keepGroupMounted) {\n      setGroupExpanded(false);\n      setIsCollapsing(true);\n      collapseTimerRef.current = window.setTimeout(() => {\n        setKeepGroupMounted(false);\n        setIsCollapsing(false);\n        collapseTimerRef.current = null;\n      }, GROUP_COLLAPSE_DURATION);\n    }\n  }, [clearGroupAnimationHandles, expanded, hasChildren, keepGroupMounted]);\n\n  React.useEffect(() => {\n    registerItem({\n      id: value,\n      value,\n      label: textLabel,\n      data: {\n        label,\n        icon,\n        endContent,\n        disabled,\n      },\n      disabled,\n      expanded,\n      hasChildren,\n      level,\n      selected,\n    });\n\n    return () => {\n      unregisterItem(value);\n    };\n  }, [\n    disabled,\n    endContent,\n    expanded,\n    hasChildren,\n    icon,\n    label,\n    level,\n    registerItem,\n    selected,\n    textLabel,\n    unregisterItem,\n    value,\n  ]);\n\n  const handleSelect = React.useCallback(() => {\n    if (disabled) {\n      return;\n    }\n\n    setSelectedId(value);\n  }, [disabled, setSelectedId, value]);\n\n  const handleToggle = React.useCallback(\n    (event: React.MouseEvent<HTMLButtonElement>) => {\n      event.stopPropagation();\n\n      if (disabled || !hasChildren) {\n        return;\n      }\n\n      toggleExpanded(value);\n    },\n    [disabled, hasChildren, toggleExpanded, value],\n  );\n\n  const handleKeyDown = React.useCallback(\n    (event: React.KeyboardEvent<HTMLDivElement>) => {\n      if (disabled) {\n        return;\n      }\n\n      if (event.key === 'Enter' || event.key === ' ') {\n        event.preventDefault();\n        handleSelect();\n        return;\n      }\n\n      if (!hasChildren) {\n        return;\n      }\n\n      if (event.key === 'ArrowRight' && !expanded) {\n        event.preventDefault();\n        toggleExpanded(value, true);\n        return;\n      }\n\n      if (event.key === 'ArrowLeft' && expanded) {\n        event.preventDefault();\n        toggleExpanded(value, false);\n      }\n    },\n    [disabled, expanded, handleSelect, hasChildren, toggleExpanded, value],\n  );\n\n  return (\n    <li\n      data-slot=\"jb-tree-item\"\n      data-value={value}\n      className={treeItemClassName}\n      {...props}\n    >\n      <div\n        data-slot=\"jb-tree-item-button\"\n        data-value={value}\n        data-has-children={hasChildren ? 'true' : undefined}\n        data-selected={selected ? 'true' : undefined}\n        data-disabled={disabled ? 'true' : undefined}\n        data-expanded={hasChildren && expanded ? 'true' : undefined}\n        data-level={level}\n        role=\"treeitem\"\n        aria-level={level}\n        aria-selected={selected}\n        aria-expanded={hasChildren ? expanded : undefined}\n        aria-disabled={disabled ? true : undefined}\n        tabIndex={disabled ? -1 : 0}\n        className={cn(treeItemButtonClassName, className)}\n        style={{ paddingLeft: `${rowPaddingLeft}px` }}\n        onClick={handleSelect}\n        onDoubleClick={\n          hasChildren && !disabled ? () => toggleExpanded(value) : undefined\n        }\n        onKeyDown={handleKeyDown}\n      >\n        {hasChildren ? (\n          <button\n            data-slot=\"jb-tree-item-disclosure\"\n            data-value={value}\n            data-expanded={expanded ? 'true' : undefined}\n            type=\"button\"\n            className={treeDisclosureClassName}\n            aria-label={\n              expanded ? `Collapse ${textLabel}` : `Expand ${textLabel}`\n            }\n            onClick={handleToggle}\n            disabled={disabled}\n          >\n            {expanded ? (\n              <ChevronDownIcon className=\"jb-tree-item-chevron\" size={16} />\n            ) : (\n              <ChevronRightIcon className=\"jb-tree-item-chevron\" size={16} />\n            )}\n          </button>\n        ) : (\n          <span\n            data-slot=\"jb-tree-item-disclosure-placeholder\"\n            aria-hidden=\"true\"\n            className={treeDisclosurePlaceholderClassName}\n          />\n        )}\n\n        <span\n          data-slot=\"jb-tree-item-content\"\n          className={treeItemContentClassName}\n        >\n          {icon ? (\n            <span\n              data-slot=\"jb-tree-item-icon\"\n              className={treeItemIconClassName}\n            >\n              {icon}\n            </span>\n          ) : null}\n\n          <span\n            data-slot=\"jb-tree-item-label\"\n            className={treeItemLabelClassName}\n          >\n            {label}\n          </span>\n\n          {endContent ? (\n            <span data-slot=\"jb-tree-item-end\" className={treeItemEndClassName}>\n              {endContent}\n            </span>\n          ) : null}\n        </span>\n      </div>\n\n      {hasChildren && keepGroupMounted ? (\n        <TreeLevelContext.Provider value={level + 1}>\n          <div\n            data-slot=\"jb-tree-group-motion\"\n            data-expanded={groupExpanded ? 'true' : undefined}\n            data-collapsing={isCollapsing ? 'true' : undefined}\n            className=\"jb-tree-group-motion\"\n            onTransitionEnd={() => {\n              if (!groupExpanded) {\n                setKeepGroupMounted(false);\n                setIsCollapsing(false);\n              }\n            }}\n          >\n            <ul\n              data-slot=\"jb-tree-group\"\n              role=\"group\"\n              aria-hidden={!expanded}\n              className={treeGroupClassName}\n            >\n              {children}\n            </ul>\n          </div>\n        </TreeLevelContext.Provider>\n      ) : null}\n    </li>\n  );\n}\n\nexport {\n  DynamicTree,\n  type DynamicTreeContextMenuIconName,\n  type DynamicTreeContextMenuItem,\n  type DynamicTreeContextMenuNode,\n  type DynamicTreeItemData,\n  type DynamicTreeProps,\n} from './dynamic-tree';\nexport { Tree, TreeItem, type TreeItemProps, type TreeProps };\n\nexport default Tree;",
+          "export * from './tree';\n\nexport { Tree as default } from './tree';",
       },
       {
-        path: 'registry/components/tree/dynamic-tree.tsx',
+        path: 'registry/components/tree/tree/index.tsx',
         type: 'registry:ui',
-        target: 'components/jetbrains-ui/tree/dynamic-tree.tsx',
+        target: 'components/jetbrains-ui/tree/tree/index.tsx',
         content:
-          "'use client';\n\nimport * as React from 'react';\nimport { flushSync } from 'react-dom';\n\nimport {\n  asyncDataLoaderFeature,\n  type ItemInstance,\n  type TreeDataLoader,\n} from '@headless-tree/core';\nimport { useTree } from '@headless-tree/react';\nimport { cn } from '@/lib/utils';\nimport {\n  AddIcon,\n  CheckIcon,\n  ChevronDownIcon,\n  ChevronRightIcon,\n  CloseIcon,\n  CopyIcon,\n  EditIcon,\n  FilterIcon,\n  RefreshIcon,\n  SearchIcon,\n  SettingsIcon,\n  TrashIcon,\n} from '@/components/jetbrains-ui/icons';\nimport {\n  ContextMenu,\n  ContextMenuContent,\n  ContextMenuItem,\n  ContextMenuSeparator,\n  ContextMenuSub,\n  ContextMenuSubContent,\n  ContextMenuSubTrigger,\n  ContextMenuTrigger,\n} from '@/components/jetbrains-ui/context-menu';\n\nexport interface DynamicTreeItemData {\n  label: string;\n  icon?: React.ReactNode;\n  endContent?: React.ReactNode;\n  disabled?: boolean;\n  isFolder?: boolean;\n}\n\nconst dynamicTreeContextMenuIcons = {\n  add: AddIcon,\n  check: CheckIcon,\n  close: CloseIcon,\n  copy: CopyIcon,\n  delete: TrashIcon,\n  edit: EditIcon,\n  filter: FilterIcon,\n  refresh: RefreshIcon,\n  search: SearchIcon,\n  settings: SettingsIcon,\n  trash: TrashIcon,\n} as const;\n\nexport type DynamicTreeContextMenuIconName =\n  keyof typeof dynamicTreeContextMenuIcons;\n\nexport interface DynamicTreeContextMenuNode {\n  id: string;\n  label: string;\n  data: DynamicTreeItemData | null;\n  item: ItemInstance<DynamicTreeItemData>;\n  disabled: boolean;\n  expanded: boolean;\n  isFolder: boolean;\n  level: number;\n  loading: boolean;\n  selected: boolean;\n}\n\nexport interface DynamicTreeContextMenuItem {\n  key: string;\n  label?: React.ReactNode;\n  text?: React.ReactNode;\n  disabled?: boolean;\n  separator?: boolean;\n  separatorAfter?: boolean;\n  iconName?: DynamicTreeContextMenuIconName;\n  variant?: 'default' | 'destructive';\n  onSelect?: (node: DynamicTreeContextMenuNode) => void;\n  children?: DynamicTreeContextMenuItem[];\n  items?: DynamicTreeContextMenuItem[];\n}\n\nexport interface DynamicTreeProps\n  extends Omit<\n    React.ComponentPropsWithoutRef<'div'>,\n    'children' | 'contextMenu'\n  > {\n  rootItemId: string;\n  dataLoader: TreeDataLoader<DynamicTreeItemData>;\n  treeLabel?: string;\n  selectedId?: string | null;\n  defaultSelectedId?: string | null;\n  onSelectedIdChange?: (id: string) => void;\n  expandedIds?: string[];\n  defaultExpandedIds?: string[];\n  onExpandedIdsChange?: (ids: string[]) => void;\n  indent?: number;\n  loadingLabel?: string;\n  loadingEndContent?: React.ReactNode;\n  contextMenu?: (\n    node: DynamicTreeContextMenuNode,\n  ) => DynamicTreeContextMenuItem[] | null | undefined;\n}\n\ntype ItemRowProps = {\n  indent: number;\n  item: ItemInstance<DynamicTreeItemData>;\n  rootItemId: string;\n  selectedId: string | null;\n  onSelect: (id: string) => void;\n  loadingLabel: string;\n  loadingEndContent: React.ReactNode;\n};\n\ntype MenuEntriesProps = {\n  items: DynamicTreeContextMenuItem[];\n  node: DynamicTreeContextMenuNode;\n};\nconst dynamicTreeRootClassName =\n  'jb-tree block w-[244px] max-w-full box-border py-2 text-[13px] leading-4 font-medium';\nconst dynamicTreeItemClassName = 'jb-tree-item m-0 p-0';\nconst dynamicTreeItemButtonClassName =\n  'jb-tree-item-button relative flex h-6 w-full min-w-0 cursor-default items-center gap-0.5 pr-4 select-none outline-none transition-[color,background-color,box-shadow] duration-150 ease-in-out';\nconst dynamicTreeDisclosureClassName =\n  'jb-tree-item-disclosure inline-flex size-4 shrink-0 items-center justify-center';\nconst dynamicTreeDisclosurePlaceholderClassName =\n  'jb-tree-item-disclosure-placeholder inline-flex size-4 shrink-0 items-center justify-center';\nconst dynamicTreeItemContentClassName =\n  'jb-tree-item-content inline-flex h-5 min-w-0 flex-1 items-center gap-1.5';\nconst dynamicTreeItemIconClassName =\n  'jb-tree-item-icon inline-flex size-4 shrink-0 items-center justify-center';\nconst dynamicTreeItemLabelClassName =\n  'jb-tree-item-label min-w-0 overflow-hidden text-ellipsis whitespace-nowrap';\nconst dynamicTreeItemEndClassName =\n  'jb-tree-item-end ml-auto pr-0.5 text-[13px] leading-4 font-medium';\n\nfunction resolveContextMenuLabel(item: DynamicTreeContextMenuItem) {\n  return item.label ?? item.text ?? item.key;\n}\n\nfunction resolveContextMenuChildren(item: DynamicTreeContextMenuItem) {\n  return item.children ?? item.items ?? [];\n}\n\nfunction resolveContextMenuSeparator(item: DynamicTreeContextMenuItem) {\n  return item.separatorAfter ?? item.separator ?? false;\n}\n\nfunction DynamicTreeContextMenuEntries({ items, node }: MenuEntriesProps) {\n  return (\n    <>\n      {items.map((item) => {\n        const label = resolveContextMenuLabel(item);\n        const childItems = resolveContextMenuChildren(item);\n        const IconComponent = item.iconName\n          ? dynamicTreeContextMenuIcons[item.iconName]\n          : null;\n        const entry = childItems.length ? (\n          <ContextMenuSub key={item.key}>\n            <ContextMenuSubTrigger disabled={item.disabled}>\n              {IconComponent ? (\n                <IconComponent\n                  aria-hidden=\"true\"\n                  className=\"size-4\"\n                  size={16}\n                />\n              ) : null}\n              <span className=\"truncate\">{label}</span>\n            </ContextMenuSubTrigger>\n            <ContextMenuSubContent>\n              <DynamicTreeContextMenuEntries items={childItems} node={node} />\n            </ContextMenuSubContent>\n          </ContextMenuSub>\n        ) : (\n          <ContextMenuItem\n            key={item.key}\n            disabled={item.disabled}\n            variant={item.variant}\n            onSelect={() => item.onSelect?.(node)}\n          >\n            {IconComponent ? (\n              <IconComponent aria-hidden=\"true\" className=\"size-4\" size={16} />\n            ) : null}\n            <span className=\"truncate\">{label}</span>\n          </ContextMenuItem>\n        );\n\n        if (!resolveContextMenuSeparator(item)) {\n          return entry;\n        }\n\n        return (\n          <React.Fragment key={`${item.key}-group`}>\n            {entry}\n            <ContextMenuSeparator />\n          </React.Fragment>\n        );\n      })}\n    </>\n  );\n}\n\nfunction DynamicTreeItemRow({\n  indent,\n  item,\n  rootItemId,\n  selectedId,\n  onSelect,\n  loadingLabel,\n  loadingEndContent,\n}: ItemRowProps) {\n  const itemProps = item.getProps() as React.ComponentPropsWithoutRef<'div'>;\n  const resolvedProps = { ...itemProps };\n  delete resolvedProps.onClick;\n  delete resolvedProps.onDoubleClick;\n  delete resolvedProps.onKeyDown;\n  const resolvedClassName = resolvedProps.className;\n  delete resolvedProps.className;\n  const itemId = item.getId();\n  const itemData = item.getItemData() as DynamicTreeItemData | null;\n  const disabled = Boolean(itemData?.disabled);\n  const selected = selectedId === itemId;\n  const expanded = item.isExpanded();\n  const folder = item.isFolder();\n  const loading = item.isLoading();\n  const level = item.getItemMeta().level;\n  const label = itemData?.label ?? loadingLabel;\n  const endContent = loading ? loadingEndContent : itemData?.endContent;\n  const rowPaddingLeft = 16 + level * indent + Math.max(0, level - 1) * 2;\n\n  const focusItem = React.useCallback(() => {\n    item.setFocused();\n  }, [item]);\n\n  const toggleFolder = React.useCallback(() => {\n    if (!folder) {\n      return;\n    }\n\n    if (expanded) {\n      item.collapse();\n    } else {\n      item.expand();\n    }\n  }, [expanded, folder, item]);\n\n  const handleSelect = React.useCallback(() => {\n    if (disabled) {\n      return;\n    }\n\n    focusItem();\n    onSelect(itemId);\n  }, [disabled, focusItem, itemId, onSelect]);\n\n  const handleKeyDown = React.useCallback(\n    (event: React.KeyboardEvent<HTMLDivElement>) => {\n      if (disabled) {\n        return;\n      }\n\n      if (event.key === 'Enter' || event.key === ' ') {\n        event.preventDefault();\n        handleSelect();\n        return;\n      }\n\n      if (event.key === 'ArrowRight') {\n        event.preventDefault();\n\n        if (folder && !expanded) {\n          item.expand();\n        } else {\n          const next = item.getItemBelow();\n          if (next) {\n            next.setFocused();\n            item.getTree().updateDomFocus();\n          }\n        }\n        return;\n      }\n\n      if (event.key === 'ArrowLeft') {\n        event.preventDefault();\n\n        if (folder && expanded) {\n          item.collapse();\n          return;\n        }\n\n        const parent = item.getParent();\n        if (parent && parent.getId() !== rootItemId) {\n          parent.setFocused();\n          item.getTree().updateDomFocus();\n        }\n        return;\n      }\n\n      if (event.key === 'ArrowDown') {\n        event.preventDefault();\n        const next = item.getItemBelow();\n        if (next) {\n          next.setFocused();\n          item.getTree().updateDomFocus();\n        }\n        return;\n      }\n\n      if (event.key === 'ArrowUp') {\n        event.preventDefault();\n        const prev = item.getItemAbove();\n        if (prev) {\n          prev.setFocused();\n          item.getTree().updateDomFocus();\n        }\n      }\n    },\n    [disabled, expanded, folder, handleSelect, item, rootItemId],\n  );\n\n  const handleDisclosureClick = React.useCallback(\n    (event: React.MouseEvent<HTMLButtonElement>) => {\n      event.stopPropagation();\n\n      if (disabled) {\n        return;\n      }\n\n      focusItem();\n      toggleFolder();\n    },\n    [disabled, focusItem, toggleFolder],\n  );\n\n  return (\n    <div\n      data-slot=\"jb-tree-item\"\n      data-value={itemId}\n      className={dynamicTreeItemClassName}\n    >\n      <div\n        {...resolvedProps}\n        data-slot=\"jb-tree-item-button\"\n        data-value={itemId}\n        data-has-children={folder ? 'true' : undefined}\n        data-selected={selected ? 'true' : undefined}\n        data-disabled={disabled ? 'true' : undefined}\n        data-expanded={folder && expanded ? 'true' : undefined}\n        data-loading={loading ? 'true' : undefined}\n        className={cn(dynamicTreeItemButtonClassName, resolvedClassName)}\n        aria-selected={selected}\n        aria-disabled={disabled ? true : undefined}\n        aria-busy={loading ? true : undefined}\n        tabIndex={disabled ? -1 : (resolvedProps.tabIndex ?? -1)}\n        style={{ paddingLeft: `${rowPaddingLeft}px` }}\n        onClick={handleSelect}\n        onDoubleClick={\n          folder && !disabled\n            ? () => {\n                focusItem();\n                toggleFolder();\n              }\n            : undefined\n        }\n        onKeyDown={handleKeyDown}\n      >\n        {folder ? (\n          <button\n            data-slot=\"jb-tree-item-disclosure\"\n            data-value={itemId}\n            data-expanded={expanded ? 'true' : undefined}\n            type=\"button\"\n            className={dynamicTreeDisclosureClassName}\n            aria-label={expanded ? `Collapse ${label}` : `Expand ${label}`}\n            onClick={handleDisclosureClick}\n            disabled={disabled}\n          >\n            {expanded ? (\n              <ChevronDownIcon className=\"jb-tree-item-chevron\" size={16} />\n            ) : (\n              <ChevronRightIcon className=\"jb-tree-item-chevron\" size={16} />\n            )}\n          </button>\n        ) : (\n          <span\n            data-slot=\"jb-tree-item-disclosure-placeholder\"\n            aria-hidden=\"true\"\n            className={dynamicTreeDisclosurePlaceholderClassName}\n          />\n        )}\n\n        <span\n          data-slot=\"jb-tree-item-content\"\n          className={dynamicTreeItemContentClassName}\n        >\n          {itemData?.icon ? (\n            <span\n              data-slot=\"jb-tree-item-icon\"\n              className={dynamicTreeItemIconClassName}\n            >\n              {itemData.icon}\n            </span>\n          ) : null}\n\n          <span\n            data-slot=\"jb-tree-item-label\"\n            className={dynamicTreeItemLabelClassName}\n          >\n            {label}\n          </span>\n\n          {endContent ? (\n            <span\n              data-slot=\"jb-tree-item-end\"\n              className={dynamicTreeItemEndClassName}\n            >\n              {endContent}\n            </span>\n          ) : null}\n        </span>\n      </div>\n    </div>\n  );\n}\n\nfunction DynamicTree({\n  rootItemId,\n  dataLoader,\n  treeLabel = 'Tree',\n  className,\n  style,\n  selectedId,\n  defaultSelectedId = null,\n  onSelectedIdChange,\n  expandedIds,\n  defaultExpandedIds = [],\n  onExpandedIdsChange,\n  indent = 16,\n  loadingLabel = 'Loading...',\n  loadingEndContent = '…',\n  contextMenu,\n  onContextMenuCapture,\n  ...props\n}: DynamicTreeProps) {\n  const [uncontrolledSelectedId, setUncontrolledSelectedId] = React.useState<\n    string | null\n  >(defaultSelectedId);\n  const [uncontrolledExpandedIds, setUncontrolledExpandedIds] =\n    React.useState<string[]>(defaultExpandedIds);\n  const [contextMenuNode, setContextMenuNode] =\n    React.useState<DynamicTreeContextMenuNode | null>(null);\n  const [contextMenuItems, setContextMenuItems] = React.useState<\n    DynamicTreeContextMenuItem[]\n  >([]);\n\n  const resolvedSelectedId =\n    selectedId !== undefined ? selectedId : uncontrolledSelectedId;\n  const resolvedExpandedIds =\n    expandedIds !== undefined ? expandedIds : uncontrolledExpandedIds;\n\n  const setSelected = React.useCallback(\n    (nextId: string) => {\n      if (selectedId === undefined) {\n        setUncontrolledSelectedId(nextId);\n      }\n\n      onSelectedIdChange?.(nextId);\n    },\n    [onSelectedIdChange, selectedId],\n  );\n\n  const setExpanded = React.useCallback(\n    (updater: string[] | ((prev: string[]) => string[])) => {\n      const nextIds =\n        typeof updater === 'function' ? updater(resolvedExpandedIds) : updater;\n\n      if (expandedIds === undefined) {\n        setUncontrolledExpandedIds(nextIds);\n      }\n\n      onExpandedIdsChange?.(nextIds);\n    },\n    [expandedIds, onExpandedIdsChange, resolvedExpandedIds],\n  );\n\n  const tree = useTree<DynamicTreeItemData>({\n    rootItemId,\n    dataLoader,\n    state: {\n      expandedItems: resolvedExpandedIds,\n    },\n    setExpandedItems: setExpanded,\n    isItemFolder: (item) => Boolean(item.getItemData()?.isFolder),\n    getItemName: (item) => item.getItemData()?.label ?? item.getId(),\n    createLoadingItemData: () => ({\n      label: loadingLabel,\n    }),\n    features: [asyncDataLoaderFeature],\n  });\n\n  const containerProps = tree.getContainerProps(\n    treeLabel,\n  ) as React.ComponentPropsWithoutRef<'div'>;\n  const items = tree.getItems();\n  const itemMap = React.useMemo(\n    () => new Map(items.map((item) => [item.getId(), item])),\n    [items],\n  );\n\n  const createContextMenuNode = React.useCallback(\n    (item: ItemInstance<DynamicTreeItemData>): DynamicTreeContextMenuNode => {\n      const itemId = item.getId();\n      const itemData = item.getItemData() as DynamicTreeItemData | null;\n\n      return {\n        id: itemId,\n        label: itemData?.label ?? itemId,\n        data: itemData,\n        item,\n        disabled: Boolean(itemData?.disabled),\n        expanded: item.isExpanded(),\n        isFolder: item.isFolder(),\n        level: item.getItemMeta().level,\n        loading: item.isLoading(),\n        selected: resolvedSelectedId === itemId,\n      };\n    },\n    [resolvedSelectedId],\n  );\n\n  const clearContextMenuState = React.useCallback(() => {\n    setContextMenuNode(null);\n    setContextMenuItems([]);\n  }, []);\n\n  const handleTreeContextMenuCapture = React.useCallback(\n    (event: React.MouseEvent<HTMLDivElement>) => {\n      onContextMenuCapture?.(event);\n\n      if (event.defaultPrevented || !contextMenu) {\n        return;\n      }\n\n      const target = event.target;\n\n      if (!(target instanceof HTMLElement)) {\n        event.preventDefault();\n        clearContextMenuState();\n        return;\n      }\n\n      const itemButton = target.closest(\n        '[data-slot=\"jb-tree-item-button\"][data-value]',\n      );\n\n      if (!(itemButton instanceof HTMLElement)) {\n        event.preventDefault();\n        clearContextMenuState();\n        return;\n      }\n\n      const itemId = itemButton.dataset.value;\n\n      if (!itemId) {\n        event.preventDefault();\n        clearContextMenuState();\n        return;\n      }\n\n      const item = itemMap.get(itemId);\n\n      if (!item) {\n        event.preventDefault();\n        clearContextMenuState();\n        return;\n      }\n\n      const nextNode = createContextMenuNode(item);\n      const nextItems = contextMenu(nextNode) ?? [];\n\n      if (nextItems.length === 0) {\n        event.preventDefault();\n        clearContextMenuState();\n        return;\n      }\n\n      flushSync(() => {\n        setContextMenuNode(nextNode);\n        setContextMenuItems(nextItems);\n      });\n    },\n    [\n      clearContextMenuState,\n      contextMenu,\n      createContextMenuNode,\n      itemMap,\n      onContextMenuCapture,\n    ],\n  );\n\n  const treeContent = (\n    <div\n      {...containerProps}\n      data-slot=\"jb-tree\"\n      className={cn(dynamicTreeRootClassName, className)}\n      style={style}\n      {...props}\n      onContextMenuCapture={\n        contextMenu ? handleTreeContextMenuCapture : onContextMenuCapture\n      }\n    >\n      {items.map((item) => (\n        <DynamicTreeItemRow\n          key={item.getKey()}\n          item={item}\n          rootItemId={rootItemId}\n          selectedId={resolvedSelectedId}\n          onSelect={setSelected}\n          indent={indent}\n          loadingLabel={loadingLabel}\n          loadingEndContent={loadingEndContent}\n        />\n      ))}\n    </div>\n  );\n\n  if (!contextMenu) {\n    return treeContent;\n  }\n\n  return (\n    <ContextMenu\n      onOpenChange={(open) => {\n        if (!open) {\n          clearContextMenuState();\n        }\n      }}\n    >\n      <ContextMenuTrigger asChild>{treeContent}</ContextMenuTrigger>\n      {contextMenuNode && contextMenuItems.length ? (\n        <ContextMenuContent>\n          <DynamicTreeContextMenuEntries\n            items={contextMenuItems}\n            node={contextMenuNode}\n          />\n        </ContextMenuContent>\n      ) : null}\n    </ContextMenu>\n  );\n}\n\nexport { DynamicTree };",
+          '\'use client\';\n\nimport * as React from \'react\';\n\nimport { cn } from \'@/lib/utils\';\nimport { ChevronDownIcon, ChevronRightIcon } from \'@/components/jetbrains-ui/icons\';\nimport { ScrollArea, ScrollViewport } from \'@/components/jetbrains-ui/scroll-area\';\nimport \'./index.css\';\n\ntype TreeContextValue = {\n  selectedId: string | null;\n  setSelectedId: (nextId: string) => void;\n  expandedIds: Set<string>;\n  toggleExpanded: (id: string, expanded?: boolean) => void;\n  indent: number;\n};\n\ninterface TreeProps extends React.ComponentPropsWithoutRef<\'ul\'> {\n  selectedId?: string | null;\n  defaultSelectedId?: string | null;\n  onSelectedIdChange?: (id: string) => void;\n  expandedIds?: string[];\n  defaultExpandedIds?: string[];\n  onExpandedIdsChange?: (ids: string[]) => void;\n  indent?: number;\n  width?: React.CSSProperties[\'width\'];\n  height?: React.CSSProperties[\'height\'];\n  scrollAreaClassName?: string;\n}\n\ninterface TreeItemProps\n  extends Omit<React.ComponentProps<\'li\'>, \'children\' | \'id\'> {\n  value: string;\n  label: React.ReactNode;\n  icon?: React.ReactNode;\n  endContent?: React.ReactNode;\n  disabled?: boolean;\n  children?: React.ReactNode;\n}\n\nconst TreeContext = React.createContext<TreeContextValue | null>(null);\nconst TreeLevelContext = React.createContext(1);\nconst GROUP_COLLAPSE_DURATION = 180;\n\ntype TreeItemButtonStyle = React.CSSProperties & {\n  \'--jb-tree-item-padding-left\': string;\n};\n\nfunction useTreeContext() {\n  const context = React.useContext(TreeContext);\n\n  if (!context) {\n    throw new Error(\'Tree components must be used within <Tree>.\');\n  }\n\n  return context;\n}\n\nfunction mergeExpandedIds(\n  current: string[],\n  id: string,\n  expanded: boolean,\n): string[] {\n  const next = new Set(current);\n\n  if (expanded) {\n    next.add(id);\n  } else {\n    next.delete(id);\n  }\n\n  return Array.from(next);\n}\n\nfunction getTreeItemLabel(label: React.ReactNode, fallback: string) {\n  if (typeof label === \'string\') {\n    return label;\n  }\n\n  return fallback;\n}\n\nfunction Tree({\n  children,\n  className,\n  style,\n  selectedId,\n  defaultSelectedId = null,\n  onSelectedIdChange,\n  expandedIds,\n  defaultExpandedIds = [],\n  onExpandedIdsChange,\n  indent = 16,\n  width,\n  height,\n  scrollAreaClassName,\n  ...props\n}: TreeProps) {\n  const [uncontrolledSelectedId, setUncontrolledSelectedId] = React.useState<\n    string | null\n  >(defaultSelectedId);\n  const [uncontrolledExpandedIds, setUncontrolledExpandedIds] =\n    React.useState<string[]>(defaultExpandedIds);\n\n  const resolvedSelectedId =\n    selectedId !== undefined ? selectedId : uncontrolledSelectedId;\n  const resolvedExpandedIds =\n    expandedIds !== undefined ? expandedIds : uncontrolledExpandedIds;\n\n  const setSelectedId = React.useCallback(\n    (nextId: string) => {\n      if (selectedId === undefined) {\n        setUncontrolledSelectedId(nextId);\n      }\n\n      onSelectedIdChange?.(nextId);\n    },\n    [onSelectedIdChange, selectedId],\n  );\n\n  const setExpandedIds = React.useCallback(\n    (nextIds: string[]) => {\n      if (expandedIds === undefined) {\n        setUncontrolledExpandedIds(nextIds);\n      }\n\n      onExpandedIdsChange?.(nextIds);\n    },\n    [expandedIds, onExpandedIdsChange],\n  );\n\n  const toggleExpanded = React.useCallback(\n    (id: string, expanded?: boolean) => {\n      const isExpanded = resolvedExpandedIds.includes(id);\n      const nextExpanded = expanded ?? !isExpanded;\n      const nextIds = mergeExpandedIds(resolvedExpandedIds, id, nextExpanded);\n\n      setExpandedIds(nextIds);\n    },\n    [resolvedExpandedIds, setExpandedIds],\n  );\n\n  const contextValue = React.useMemo<TreeContextValue>(\n    () => ({\n      selectedId: resolvedSelectedId,\n      setSelectedId,\n      expandedIds: new Set(resolvedExpandedIds),\n      toggleExpanded,\n      indent,\n    }),\n    [\n      indent,\n      resolvedExpandedIds,\n      resolvedSelectedId,\n      setSelectedId,\n      toggleExpanded,\n    ],\n  );\n\n  const treeList = (\n    <ul\n      data-slot="jb-tree"\n      role="tree"\n      className={cn(\n        \'jb-tree jb-static-tree m-0 list-none p-0 block w-full min-w-[244px] box-border py-2 text-[13px] leading-4 font-medium [font-family:var(--jb-font-sans),sans-serif]\',\n        className,\n      )}\n      style={style}\n      {...props}\n    >\n      <TreeLevelContext.Provider value={1}>\n        {children}\n      </TreeLevelContext.Provider>\n    </ul>\n  );\n\n  return (\n    <TreeContext.Provider value={contextValue}>\n      <ScrollArea\n        data-slot="jb-tree-scroll-area"\n        className={cn(\'max-w-full\', scrollAreaClassName)}\n        style={{ width, height }}\n      >\n        <ScrollViewport\n          data-slot="jb-tree-scroll-viewport"\n          className="size-full"\n        >\n          {treeList}\n        </ScrollViewport>\n      </ScrollArea>\n    </TreeContext.Provider>\n  );\n}\n\nfunction TreeItem({\n  className,\n  value,\n  label,\n  icon,\n  endContent,\n  disabled = false,\n  children,\n  ...props\n}: TreeItemProps) {\n  const { selectedId, setSelectedId, expandedIds, toggleExpanded, indent } =\n    useTreeContext();\n  const level = React.useContext(TreeLevelContext);\n\n  const hasChildren = React.Children.count(children) > 0;\n  const expanded = hasChildren ? expandedIds.has(value) : false;\n  const selected = selectedId === value;\n  const textLabel = getTreeItemLabel(label, value);\n  const [keepGroupMounted, setKeepGroupMounted] = React.useState(expanded);\n  const [groupExpanded, setGroupExpanded] = React.useState(expanded);\n  const [isCollapsing, setIsCollapsing] = React.useState(false);\n  const hasMountedRef = React.useRef(false);\n  const collapseTimerRef = React.useRef<number | null>(null);\n  const expandRafRef = React.useRef<number | null>(null);\n\n  const rowPaddingLeft = 16 + (level - 1) * indent + Math.max(0, level - 2) * 2;\n  const resolvedItemButtonStyle: TreeItemButtonStyle = {\n    \'--jb-tree-item-padding-left\': `${rowPaddingLeft}px`,\n  };\n\n  const clearGroupAnimationHandles = React.useCallback(() => {\n    if (collapseTimerRef.current !== null) {\n      window.clearTimeout(collapseTimerRef.current);\n      collapseTimerRef.current = null;\n    }\n\n    if (expandRafRef.current !== null) {\n      window.cancelAnimationFrame(expandRafRef.current);\n      expandRafRef.current = null;\n    }\n  }, []);\n\n  React.useEffect(() => {\n    return () => {\n      clearGroupAnimationHandles();\n    };\n  }, [clearGroupAnimationHandles]);\n\n  React.useEffect(() => {\n    clearGroupAnimationHandles();\n\n    if (!hasChildren) {\n      setKeepGroupMounted(false);\n      setGroupExpanded(false);\n      setIsCollapsing(false);\n      return;\n    }\n\n    if (!hasMountedRef.current) {\n      hasMountedRef.current = true;\n      setKeepGroupMounted(expanded);\n      setGroupExpanded(expanded);\n      setIsCollapsing(false);\n      return;\n    }\n\n    if (expanded) {\n      setKeepGroupMounted(true);\n      setIsCollapsing(false);\n\n      // Mount collapsed first, then expand on next frame so transition is visible.\n      setGroupExpanded(false);\n      expandRafRef.current = window.requestAnimationFrame(() => {\n        expandRafRef.current = window.requestAnimationFrame(() => {\n          setGroupExpanded(true);\n          expandRafRef.current = null;\n        });\n      });\n\n      return;\n    }\n\n    if (keepGroupMounted) {\n      setGroupExpanded(false);\n      setIsCollapsing(true);\n      collapseTimerRef.current = window.setTimeout(() => {\n        setKeepGroupMounted(false);\n        setIsCollapsing(false);\n        collapseTimerRef.current = null;\n      }, GROUP_COLLAPSE_DURATION);\n    }\n  }, [clearGroupAnimationHandles, expanded, hasChildren, keepGroupMounted]);\n\n  const handleSelect = React.useCallback(() => {\n    if (disabled) {\n      return;\n    }\n\n    setSelectedId(value);\n  }, [disabled, setSelectedId, value]);\n\n  const handleToggle = React.useCallback(\n    (event: React.MouseEvent<HTMLButtonElement>) => {\n      event.stopPropagation();\n\n      if (disabled || !hasChildren) {\n        return;\n      }\n\n      toggleExpanded(value);\n    },\n    [disabled, hasChildren, toggleExpanded, value],\n  );\n\n  const handleKeyDown = React.useCallback(\n    (event: React.KeyboardEvent<HTMLDivElement>) => {\n      if (disabled) {\n        return;\n      }\n\n      if (event.key === \'Enter\' || event.key === \' \') {\n        event.preventDefault();\n        handleSelect();\n        return;\n      }\n\n      if (!hasChildren) {\n        return;\n      }\n\n      if (event.key === \'ArrowRight\' && !expanded) {\n        event.preventDefault();\n        toggleExpanded(value, true);\n        return;\n      }\n\n      if (event.key === \'ArrowLeft\' && expanded) {\n        event.preventDefault();\n        toggleExpanded(value, false);\n      }\n    },\n    [disabled, expanded, handleSelect, hasChildren, toggleExpanded, value],\n  );\n\n  return (\n    <li\n      data-slot="jb-tree-item"\n      data-value={value}\n      className="jb-tree-item m-0 list-none p-0"\n      {...props}\n    >\n      <div\n        data-slot="jb-tree-item-button"\n        data-value={value}\n        data-has-children={hasChildren ? \'true\' : undefined}\n        data-selected={selected ? \'true\' : undefined}\n        data-disabled={disabled ? \'true\' : undefined}\n        data-expanded={hasChildren && expanded ? \'true\' : undefined}\n        data-level={level}\n        role="treeitem"\n        aria-level={level}\n        aria-selected={selected}\n        aria-expanded={hasChildren ? expanded : undefined}\n        aria-disabled={disabled ? true : undefined}\n        tabIndex={disabled ? -1 : 0}\n        className={cn(\n          \'jb-tree-item-button relative box-border flex h-6 w-full min-w-0 cursor-default items-center gap-0.5 pr-4 pl-[var(--jb-tree-item-padding-left)] select-none outline-none transition-[color] duration-150 ease-in-out\',\n          className,\n        )}\n        style={resolvedItemButtonStyle}\n        onClick={handleSelect}\n        onDoubleClick={\n          hasChildren && !disabled ? () => toggleExpanded(value) : undefined\n        }\n        onKeyDown={handleKeyDown}\n      >\n        <span\n          aria-hidden="true"\n          data-slot="jb-tree-item-overlay"\n          className="jb-tree-item-overlay pointer-events-none absolute inset-y-0 left-3 right-3 rounded-[4px] transition-[background-color,box-shadow] duration-150 ease-in-out"\n        />\n\n        {hasChildren ? (\n          <button\n            data-slot="jb-tree-item-disclosure"\n            data-value={value}\n            data-expanded={expanded ? \'true\' : undefined}\n            type="button"\n            className="jb-tree-item-disclosure relative z-10 inline-flex size-4 shrink-0 items-center justify-center border-0 bg-transparent p-0 outline-none"\n            aria-label={\n              expanded ? `Collapse ${textLabel}` : `Expand ${textLabel}`\n            }\n            onClick={handleToggle}\n            disabled={disabled}\n          >\n            {expanded ? (\n              <ChevronDownIcon className="size-4 scale-[0.78]" size={16} />\n            ) : (\n              <ChevronRightIcon className="size-4 scale-[0.78]" size={16} />\n            )}\n          </button>\n        ) : (\n          <span\n            data-slot="jb-tree-item-disclosure-placeholder"\n            aria-hidden="true"\n            className="jb-tree-item-disclosure-placeholder relative z-10 inline-flex size-4 shrink-0 items-center justify-center"\n          />\n        )}\n\n        <span\n          data-slot="jb-tree-item-content"\n          className="jb-tree-item-content relative z-10 inline-flex h-5 min-w-0 flex-1 items-center gap-1.5"\n        >\n          {icon ? (\n            <span\n              data-slot="jb-tree-item-icon"\n              className="jb-tree-item-icon inline-flex size-4 shrink-0 items-center justify-center [&>svg]:size-4"\n            >\n              {icon}\n            </span>\n          ) : null}\n\n          <span\n            data-slot="jb-tree-item-label"\n            className="jb-tree-item-label min-w-0 overflow-hidden text-ellipsis whitespace-nowrap"\n          >\n            {label}\n          </span>\n\n          {endContent ? (\n            <span\n              data-slot="jb-tree-item-end"\n              className="jb-tree-item-end ml-auto pr-0.5 text-[13px] leading-4 font-medium"\n            >\n              {endContent}\n            </span>\n          ) : null}\n        </span>\n      </div>\n\n      {hasChildren && keepGroupMounted ? (\n        <TreeLevelContext.Provider value={level + 1}>\n          <div\n            data-slot="jb-tree-group-motion"\n            data-expanded={groupExpanded ? \'true\' : undefined}\n            data-collapsing={isCollapsing ? \'true\' : undefined}\n            className="jb-tree-group-motion"\n            onTransitionEnd={() => {\n              if (!groupExpanded) {\n                setKeepGroupMounted(false);\n                setIsCollapsing(false);\n              }\n            }}\n          >\n            <ul\n              data-slot="jb-tree-group"\n              role="group"\n              aria-hidden={!expanded}\n              className="jb-tree-group m-0 list-none p-0"\n            >\n              {children}\n            </ul>\n          </div>\n        </TreeLevelContext.Provider>\n      ) : null}\n    </li>\n  );\n}\n\nexport { Tree, TreeItem, type TreeItemProps, type TreeProps };\n\nexport default Tree;',
       },
       {
-        path: 'registry/components/tree/index.css',
+        path: 'registry/components/tree/tree/index.css',
         type: 'registry:style',
-        target: 'components/jetbrains-ui/tree/index.css',
+        target: 'components/jetbrains-ui/tree/tree/index.css',
         content:
-          ":root {\n  --jb-tree-bg: transparent;\n  --jb-tree-text: var(--jb-gray-1);\n  --jb-tree-muted: var(--jb-gray-7);\n  --jb-tree-icon: var(--jb-gray-6);\n  --jb-tree-disclosure: var(--jb-gray-7);\n  --jb-tree-disclosure-hover: var(--jb-gray-5);\n  --jb-tree-item-hover-bg: var(--jb-gray-12);\n  --jb-tree-item-active-bg: var(--jb-gray-11);\n  --jb-tree-item-selected-bg: var(--jb-blue-11);\n  --jb-tree-item-selected-text: var(--jb-gray-1);\n  --jb-tree-focus-ring: var(--jb-blue-4);\n  --jb-tree-focus-ring-offset: var(--jb-gray-14);\n  --jb-tree-disabled-text: var(--jb-gray-8);\n}\n\n.dark {\n  --jb-tree-bg: transparent;\n  --jb-tree-text: var(--jb-gray-12);\n  --jb-tree-muted: var(--jb-gray-7);\n  --jb-tree-icon: var(--jb-gray-10);\n  --jb-tree-disclosure: var(--jb-gray-10);\n  --jb-tree-disclosure-hover: var(--jb-gray-12);\n  --jb-tree-item-hover-bg: var(--jb-gray-3);\n  --jb-tree-item-active-bg: var(--jb-gray-4);\n  --jb-tree-item-selected-bg: var(--jb-blue-2);\n  --jb-tree-item-selected-text: var(--jb-gray-12);\n  --jb-tree-focus-ring: var(--jb-blue-6);\n  --jb-tree-focus-ring-offset: var(--jb-gray-2);\n  --jb-tree-disabled-text: var(--jb-gray-7);\n}\n\n.jb-tree {\n  background: var(--jb-tree-bg);\n  color: var(--jb-tree-text);\n  font-family: var(--jb-font-sans), sans-serif;\n}\n\n.jb-tree-group-motion {\n  display: grid;\n  grid-template-rows: 0fr;\n  min-height: 0;\n  opacity: 0;\n  transform: translateY(-2px);\n  pointer-events: none;\n  transition:\n    grid-template-rows 170ms var(--jb-ease-out),\n    opacity 140ms var(--jb-ease-out),\n    transform 170ms var(--jb-ease-out);\n}\n\n.jb-tree-group-motion[data-expanded='true'] {\n  grid-template-rows: 1fr;\n  opacity: 1;\n  transform: translateY(0);\n  pointer-events: auto;\n}\n\n.jb-tree-group-motion .jb-tree-group {\n  min-height: 0;\n  overflow: hidden;\n}\n\n.jb-tree-item-button {\n  position: relative;\n  color: var(--jb-tree-text);\n  box-sizing: border-box;\n}\n\n.jb-tree-item-button::before {\n  content: '';\n  position: absolute;\n  inset: 0 12px;\n  border-radius: 4px;\n  background: transparent;\n  transition:\n    background-color var(--jb-transition-fast) ease-in-out,\n    box-shadow var(--jb-transition-fast) ease-in-out;\n}\n\n.jb-tree-item-button > * {\n  position: relative;\n  z-index: 1;\n}\n\n.jb-tree-item-button:hover:not([data-disabled='true']) {\n  background: transparent;\n}\n\n.jb-tree-item-button:hover:not([data-disabled='true'])::before {\n  background: var(--jb-tree-item-hover-bg);\n}\n\n.jb-tree-item-button:active:not([data-disabled='true']) {\n  background: transparent;\n}\n\n.jb-tree-item-button:active:not([data-disabled='true'])::before {\n  background: var(--jb-tree-item-active-bg);\n}\n\n.jb-tree-item-button[data-selected='true'] {\n  background: transparent;\n  color: var(--jb-tree-item-selected-text);\n}\n\n.jb-tree-item-button[data-selected='true']::before {\n  background: var(--jb-tree-item-selected-bg);\n}\n\n.jb-tree-item-button:focus-visible::before {\n  box-shadow:\n    0 0 0 1px var(--jb-tree-focus-ring-offset),\n    0 0 0 3px var(--jb-tree-focus-ring);\n}\n\n.jb-tree-item-button[data-disabled='true'] {\n  color: var(--jb-tree-disabled-text);\n  cursor: not-allowed;\n}\n\n.jb-tree-item-disclosure {\n  border: 0;\n  background: transparent;\n  padding: 0;\n  color: var(--jb-tree-disclosure);\n  outline: none;\n}\n\n.jb-tree-item-disclosure:hover:not(:disabled) {\n  color: var(--jb-tree-disclosure-hover);\n}\n\n.jb-tree-item-disclosure:disabled {\n  cursor: not-allowed;\n}\n\n.jb-tree-item-chevron {\n  width: 16px;\n  height: 16px;\n  transform: scale(0.78);\n  transform-origin: center;\n}\n\n.jb-tree-item-icon {\n  color: var(--jb-tree-icon);\n}\n\n.jb-tree-item-button[data-selected='true'] .jb-tree-item-icon,\n.jb-tree-item-button[data-selected='true'] .jb-tree-item-disclosure {\n  color: currentColor;\n}\n\n.jb-tree-item-icon > svg {\n  width: 16px;\n  height: 16px;\n}\n\n.jb-tree-item-end {\n  color: var(--jb-tree-muted);\n}",
+          ":root {\n  --jb-tree-bg: transparent;\n  --jb-tree-text: var(--jb-gray-1);\n  --jb-tree-muted: var(--jb-gray-7);\n  --jb-tree-icon: var(--jb-gray-6);\n  --jb-tree-disclosure: var(--jb-gray-7);\n  --jb-tree-disclosure-hover: var(--jb-gray-5);\n  --jb-tree-item-hover-bg: var(--jb-gray-12);\n  --jb-tree-item-active-bg: var(--jb-gray-11);\n  --jb-tree-item-selected-bg: var(--jb-blue-11);\n  --jb-tree-item-selected-text: var(--jb-gray-1);\n  --jb-tree-focus-ring: var(--jb-blue-4);\n  --jb-tree-focus-ring-offset: var(--jb-gray-14);\n  --jb-tree-disabled-text: var(--jb-gray-8);\n}\n\n.dark {\n  --jb-tree-bg: transparent;\n  --jb-tree-text: var(--jb-gray-12);\n  --jb-tree-muted: var(--jb-gray-7);\n  --jb-tree-icon: var(--jb-gray-10);\n  --jb-tree-disclosure: var(--jb-gray-10);\n  --jb-tree-disclosure-hover: var(--jb-gray-12);\n  --jb-tree-item-hover-bg: var(--jb-gray-3);\n  --jb-tree-item-active-bg: var(--jb-gray-4);\n  --jb-tree-item-selected-bg: var(--jb-blue-2);\n  --jb-tree-item-selected-text: var(--jb-gray-12);\n  --jb-tree-focus-ring: var(--jb-blue-6);\n  --jb-tree-focus-ring-offset: var(--jb-gray-2);\n  --jb-tree-disabled-text: var(--jb-gray-7);\n}\n\n.jb-static-tree {\n  background: var(--jb-tree-bg);\n  color: var(--jb-tree-text);\n  font-family: var(--jb-font-sans), sans-serif;\n}\n\n.jb-static-tree .jb-tree-group-motion {\n  display: grid;\n  grid-template-rows: 0fr;\n  min-height: 0;\n  opacity: 0;\n  transform: translateY(-2px);\n  pointer-events: none;\n  transition:\n    grid-template-rows 170ms var(--jb-ease-out),\n    opacity 140ms var(--jb-ease-out),\n    transform 170ms var(--jb-ease-out);\n}\n\n.jb-static-tree .jb-tree-group-motion[data-expanded='true'] {\n  grid-template-rows: 1fr;\n  opacity: 1;\n  transform: translateY(0);\n  pointer-events: auto;\n}\n\n.jb-static-tree .jb-tree-group-motion .jb-tree-group {\n  min-height: 0;\n  overflow: hidden;\n}\n\n.jb-static-tree .jb-tree-item-button {\n  position: relative;\n  color: var(--jb-tree-text);\n  box-sizing: border-box;\n}\n\n.jb-static-tree .jb-tree-item-overlay {\n  background-color: transparent;\n}\n\n.jb-static-tree .jb-tree-item-button:hover:not([data-disabled='true']) {\n  background: transparent;\n}\n\n.jb-static-tree\n  .jb-tree-item-button:hover:not([data-disabled='true']):not(\n    [data-selected='true']\n  )\n  .jb-tree-item-overlay {\n  background: var(--jb-tree-item-hover-bg);\n}\n\n.jb-static-tree .jb-tree-item-button:active:not([data-disabled='true']) {\n  background: transparent;\n}\n\n.jb-static-tree\n  .jb-tree-item-button:active:not([data-disabled='true']):not(\n    [data-selected='true']\n  )\n  .jb-tree-item-overlay {\n  background: var(--jb-tree-item-active-bg);\n}\n\n.jb-static-tree .jb-tree-item-button[data-selected='true'] {\n  background: transparent;\n  color: var(--jb-tree-item-selected-text);\n}\n\n.jb-static-tree\n  .jb-tree-item-button[data-selected='true']\n  .jb-tree-item-overlay {\n  background: var(--jb-tree-item-selected-bg);\n}\n\n.jb-static-tree\n  .jb-tree-item-button:focus-visible:not([data-disabled='true'])\n  .jb-tree-item-overlay {\n  box-shadow:\n    0 0 0 1px var(--jb-tree-focus-ring-offset),\n    0 0 0 3px var(--jb-tree-focus-ring);\n}\n\n.jb-static-tree .jb-tree-item-button[data-disabled='true'] {\n  color: var(--jb-tree-disabled-text);\n  cursor: not-allowed;\n}\n\n.jb-static-tree .jb-tree-item-disclosure {\n  border: 0;\n  background: transparent;\n  padding: 0;\n  color: var(--jb-tree-disclosure);\n  outline: none;\n}\n\n.jb-static-tree .jb-tree-item-disclosure:disabled {\n  cursor: not-allowed;\n}\n\n.jb-static-tree\n  .jb-tree-item-button:not([data-selected='true']):not([data-disabled='true'])\n  .jb-tree-item-disclosure:hover {\n  color: var(--jb-tree-disclosure-hover);\n}\n\n.jb-static-tree .jb-tree-item-icon {\n  color: var(--jb-tree-icon);\n}\n\n.jb-static-tree\n  .jb-tree-item-button[data-disabled='true']\n  .jb-tree-item-disclosure {\n  color: var(--jb-tree-disabled-text);\n}\n\n.jb-static-tree .jb-tree-item-button[data-selected='true'] .jb-tree-item-icon,\n.jb-static-tree\n  .jb-tree-item-button[data-selected='true']\n  .jb-tree-item-disclosure {\n  color: currentColor;\n}\n\n.jb-static-tree .jb-tree-item-end {\n  color: var(--jb-tree-muted);\n}",
+      },
+      {
+        path: 'registry/components/tree/dynamic-tree/index.tsx',
+        type: 'registry:ui',
+        target: 'components/jetbrains-ui/tree/dynamic-tree/index.tsx',
+        content:
+          '\'use client\';\n\nimport * as React from \'react\';\n\nimport {\n  asyncDataLoaderFeature,\n  hotkeysCoreFeature,\n  selectionFeature,\n  type ItemInstance,\n  type TreeDataLoader,\n} from \'@headless-tree/core\';\nimport { useTree } from \'@headless-tree/react\';\nimport {\n  ChevronDownIcon,\n  ChevronRightIcon,\n  LoaderIcon,\n} from \'@/components/jetbrains-ui/icons\';\nimport { ScrollArea, ScrollViewport } from \'@/components/jetbrains-ui/scroll-area\';\nimport \'./index.css\';\n\nconst DYNAMIC_TREE_ROOT_ITEM_ID = \'__jb-dynamic-tree-root__\';\n\nexport interface DynamicTreeItemData {\n  label: string;\n  icon?: React.ReactNode;\n  endContent?: React.ReactNode;\n  disabled?: boolean;\n  isFolder?: boolean;\n}\n\nexport interface DynamicTreeLoadedItem<\n  TItem extends DynamicTreeItemData = DynamicTreeItemData,\n> {\n  id: string;\n  data: TItem;\n}\n\nexport type DynamicTreeLoadData<\n  TItem extends DynamicTreeItemData = DynamicTreeItemData,\n> = (\n  itemId: string | null,\n) => DynamicTreeLoadedItem<TItem>[] | Promise<DynamicTreeLoadedItem<TItem>[]>;\n\nexport interface DynamicTreeProps<\n  TItem extends DynamicTreeItemData = DynamicTreeItemData,\n> extends Omit<\n    React.ComponentPropsWithoutRef<\'div\'>,\n    \'children\' | \'className\' | \'style\'\n  > {\n  loadData: DynamicTreeLoadData<TItem>;\n  treeLabel?: string;\n  indent?: number;\n  width?: React.CSSProperties[\'width\'];\n  height?: React.CSSProperties[\'height\'];\n  loadingLabel?: string;\n}\n\ntype ItemRowProps<TItem extends DynamicTreeItemData> = {\n  indent: number;\n  item: ItemInstance<TItem>;\n  loadingLabel: string;\n};\n\ntype DynamicTreeItemButtonStyle = React.CSSProperties & {\n  \'--jb-tree-item-padding-left\': string;\n};\n\nfunction DynamicTreeItemRow<TItem extends DynamicTreeItemData>({\n  indent,\n  item,\n  loadingLabel,\n}: ItemRowProps<TItem>) {\n  const {\n    onClick: ignoredOnClick,\n    onDoubleClick: ignoredOnDoubleClick,\n    onKeyDown: resolvedOnKeyDown,\n    style: resolvedStyle,\n    tabIndex: resolvedTabIndex,\n    ...resolvedProps\n  } = item.getProps() as React.ComponentPropsWithoutRef<\'div\'>;\n  void ignoredOnClick;\n  void ignoredOnDoubleClick;\n\n  return (\n    <div\n      data-slot="jb-tree-item"\n      data-value={item.getId()}\n      className="jb-tree-item m-0 p-0"\n    >\n      <div\n        {...resolvedProps}\n        data-slot="jb-tree-item-button"\n        data-value={item.getId()}\n        data-has-children={item.isFolder() ? \'true\' : undefined}\n        data-selected={item.isSelected() ? \'true\' : undefined}\n        data-disabled={\n          (item.getItemData() as TItem | null)?.disabled ? \'true\' : undefined\n        }\n        data-expanded={\n          item.isFolder() && item.isExpanded() ? \'true\' : undefined\n        }\n        data-loading={item.isLoading() ? \'true\' : undefined}\n        aria-selected={item.isSelected()}\n        aria-disabled={\n          (item.getItemData() as TItem | null)?.disabled ? true : undefined\n        }\n        aria-busy={item.isLoading() ? true : undefined}\n        tabIndex={\n          (item.getItemData() as TItem | null)?.disabled\n            ? -1\n            : (resolvedTabIndex ?? -1)\n        }\n        style={\n          {\n            ...resolvedStyle,\n            \'--jb-tree-item-padding-left\': `${\n              16 +\n              item.getItemMeta().level * indent +\n              Math.max(0, item.getItemMeta().level - 1) * 2\n            }px`,\n          } as DynamicTreeItemButtonStyle\n        }\n        onClick={() => {\n          if ((item.getItemData() as TItem | null)?.disabled) {\n            return;\n          }\n\n          item.setFocused();\n          item.select();\n        }}\n        onDoubleClick={\n          item.isFolder() && !(item.getItemData() as TItem | null)?.disabled\n            ? () => {\n                item.setFocused();\n\n                if (item.isExpanded()) {\n                  item.collapse();\n                } else {\n                  item.expand();\n                }\n              }\n            : undefined\n        }\n        onKeyDown={(event) => {\n          if ((item.getItemData() as TItem | null)?.disabled) {\n            resolvedOnKeyDown?.(event);\n            return;\n          }\n\n          if (event.key === \'Enter\' || event.key === \' \') {\n            event.preventDefault();\n            item.setFocused();\n            item.select();\n            return;\n          }\n\n          resolvedOnKeyDown?.(event);\n        }}\n        className="jb-tree-item-button relative box-border flex h-6 w-full min-w-0 cursor-default items-center gap-0.5 pr-4 pl-[var(--jb-tree-item-padding-left)] select-none outline-none transition-[color] duration-150 ease-in-out data-[disabled=true]:cursor-not-allowed"\n      >\n        <span\n          aria-hidden="true"\n          data-slot="jb-tree-item-overlay"\n          className="jb-tree-item-overlay pointer-events-none absolute inset-y-0 left-3 right-3 rounded-[4px] transition-[background-color,box-shadow] duration-150 ease-in-out"\n        />\n\n        {item.isFolder() ? (\n          <button\n            data-slot="jb-tree-item-disclosure"\n            data-value={item.getId()}\n            data-expanded={item.isExpanded() ? \'true\' : undefined}\n            data-loading={item.isLoading() ? \'true\' : undefined}\n            type="button"\n            className="jb-tree-item-disclosure relative z-10 inline-flex size-4 shrink-0 items-center justify-center border-0 bg-transparent p-0 outline-none disabled:cursor-not-allowed"\n            aria-label={\n              item.isLoading()\n                ? `Loading ${(item.getItemData() as TItem | null)?.label ?? loadingLabel}`\n                : item.isExpanded()\n                  ? `Collapse ${(item.getItemData() as TItem | null)?.label ?? loadingLabel}`\n                  : `Expand ${(item.getItemData() as TItem | null)?.label ?? loadingLabel}`\n            }\n            onClick={(event) => {\n              event.stopPropagation();\n\n              if ((item.getItemData() as TItem | null)?.disabled) {\n                return;\n              }\n\n              item.setFocused();\n\n              if (item.isExpanded()) {\n                item.collapse();\n              } else {\n                item.expand();\n              }\n            }}\n            disabled={Boolean((item.getItemData() as TItem | null)?.disabled)}\n          >\n            {item.isLoading() ? (\n              <LoaderIcon\n                className="size-4 [animation-duration:900ms]"\n                size={16}\n              />\n            ) : item.isExpanded() ? (\n              <ChevronDownIcon className="size-4 scale-[0.78]" size={16} />\n            ) : (\n              <ChevronRightIcon className="size-4 scale-[0.78]" size={16} />\n            )}\n          </button>\n        ) : (\n          <span\n            data-slot="jb-tree-item-disclosure-placeholder"\n            aria-hidden="true"\n            className="jb-tree-item-disclosure-placeholder relative z-10 inline-flex size-4 shrink-0 items-center justify-center"\n          />\n        )}\n\n        <span\n          data-slot="jb-tree-item-content"\n          className="jb-tree-item-content relative z-10 inline-flex h-5 min-w-0 flex-1 items-center gap-1.5"\n        >\n          {(item.getItemData() as TItem | null)?.icon ? (\n            <span\n              data-slot="jb-tree-item-icon"\n              className="jb-tree-item-icon inline-flex size-4 shrink-0 items-center justify-center [&>svg]:size-4"\n            >\n              {(item.getItemData() as TItem | null)?.icon}\n            </span>\n          ) : null}\n\n          <span\n            data-slot="jb-tree-item-label"\n            className="jb-tree-item-label min-w-0 overflow-hidden text-ellipsis whitespace-nowrap"\n          >\n            {(item.getItemData() as TItem | null)?.label ?? loadingLabel}\n          </span>\n\n          {(item.getItemData() as TItem | null)?.endContent ? (\n            <span\n              data-slot="jb-tree-item-end"\n              className="jb-tree-item-end ml-auto pr-0.5 text-[13px] leading-4 font-medium"\n            >\n              {(item.getItemData() as TItem | null)?.endContent}\n            </span>\n          ) : null}\n        </span>\n      </div>\n    </div>\n  );\n}\n\nfunction DynamicTree<TItem extends DynamicTreeItemData>({\n  loadData,\n  treeLabel = \'Tree\',\n  indent = 16,\n  width,\n  height,\n  loadingLabel = \'Loading...\',\n  ...props\n}: DynamicTreeProps<TItem>) {\n  const loadedItemsRef = React.useRef<Record<string, TItem>>({});\n\n  const dataLoader = React.useMemo<TreeDataLoader<TItem>>(\n    () => ({\n      getItem: async (itemId: string) => {\n        if (itemId === DYNAMIC_TREE_ROOT_ITEM_ID) {\n          return {\n            label: treeLabel,\n            isFolder: true,\n          } as TItem;\n        }\n\n        return (\n          loadedItemsRef.current[itemId] ??\n          ({\n            label: itemId,\n          } as TItem)\n        );\n      },\n      getChildrenWithData: async (itemId: string) =>\n        loadData(itemId === DYNAMIC_TREE_ROOT_ITEM_ID ? null : itemId),\n    }),\n    [loadData, treeLabel],\n  );\n\n  const tree = useTree<TItem>({\n    rootItemId: DYNAMIC_TREE_ROOT_ITEM_ID,\n    dataLoader,\n    isItemFolder: (item) => Boolean(item.getItemData()?.isFolder),\n    getItemName: (item) => item.getItemData()?.label ?? item.getId(),\n    onLoadedItem: (itemId, item) => {\n      loadedItemsRef.current[itemId] = item;\n    },\n    createLoadingItemData: () =>\n      ({\n        label: loadingLabel,\n      }) as TItem,\n    features: [asyncDataLoaderFeature, selectionFeature, hotkeysCoreFeature],\n  });\n\n  const containerProps = tree.getContainerProps(\n    treeLabel,\n  ) as React.ComponentPropsWithoutRef<\'div\'>;\n  const items = tree.getItems();\n\n  return (\n    <ScrollArea\n      data-slot="jb-tree-scroll-area"\n      className="max-w-full"\n      style={{ width, height }}\n    >\n      <ScrollViewport data-slot="jb-tree-scroll-viewport" className="size-full">\n        <div\n          {...containerProps}\n          data-slot="jb-tree"\n          className="jb-tree jb-dynamic-tree block w-full min-w-[244px] box-border py-2 text-[13px] leading-4 font-medium [font-family:var(--jb-font-sans),sans-serif]"\n          {...props}\n        >\n          {items.map((item) => (\n            <DynamicTreeItemRow\n              key={item.getKey()}\n              item={item}\n              indent={indent}\n              loadingLabel={loadingLabel}\n            />\n          ))}\n        </div>\n      </ScrollViewport>\n    </ScrollArea>\n  );\n}\n\nexport { DynamicTree };\n\nexport default DynamicTree;',
+      },
+      {
+        path: 'registry/components/tree/dynamic-tree/index.css',
+        type: 'registry:style',
+        target: 'components/jetbrains-ui/tree/dynamic-tree/index.css',
+        content:
+          ":root {\n  --jb-tree-bg: transparent;\n  --jb-tree-text: var(--jb-gray-1);\n  --jb-tree-muted: var(--jb-gray-7);\n  --jb-tree-icon: var(--jb-gray-6);\n  --jb-tree-disclosure: var(--jb-gray-7);\n  --jb-tree-disclosure-hover: var(--jb-gray-5);\n  --jb-tree-item-hover-bg: var(--jb-gray-12);\n  --jb-tree-item-active-bg: var(--jb-gray-11);\n  --jb-tree-item-selected-bg: var(--jb-blue-11);\n  --jb-tree-item-selected-text: var(--jb-gray-1);\n  --jb-tree-focus-ring: var(--jb-blue-4);\n  --jb-tree-focus-ring-offset: var(--jb-gray-14);\n  --jb-tree-disabled-text: var(--jb-gray-8);\n}\n\n.dark {\n  --jb-tree-bg: transparent;\n  --jb-tree-text: var(--jb-gray-12);\n  --jb-tree-muted: var(--jb-gray-7);\n  --jb-tree-icon: var(--jb-gray-10);\n  --jb-tree-disclosure: var(--jb-gray-10);\n  --jb-tree-disclosure-hover: var(--jb-gray-12);\n  --jb-tree-item-hover-bg: var(--jb-gray-3);\n  --jb-tree-item-active-bg: var(--jb-gray-4);\n  --jb-tree-item-selected-bg: var(--jb-blue-2);\n  --jb-tree-item-selected-text: var(--jb-gray-12);\n  --jb-tree-focus-ring: var(--jb-blue-6);\n  --jb-tree-focus-ring-offset: var(--jb-gray-2);\n  --jb-tree-disabled-text: var(--jb-gray-7);\n}\n\n.jb-dynamic-tree {\n  background-color: var(--jb-tree-bg);\n  color: var(--jb-tree-text);\n}\n\n.jb-dynamic-tree .jb-tree-item-button {\n  color: var(--jb-tree-text);\n}\n\n.jb-dynamic-tree .jb-tree-item-button[data-disabled='true'] {\n  color: var(--jb-tree-disabled-text);\n}\n\n.jb-dynamic-tree .jb-tree-item-button[data-selected='true'] {\n  color: var(--jb-tree-item-selected-text);\n}\n\n.jb-dynamic-tree .jb-tree-item-overlay {\n  background-color: transparent;\n}\n\n.jb-dynamic-tree\n  .jb-tree-item-button:hover:not([data-disabled='true']):not(\n    [data-selected='true']\n  )\n  .jb-tree-item-overlay {\n  background-color: var(--jb-tree-item-hover-bg);\n}\n\n.jb-dynamic-tree\n  .jb-tree-item-button:active:not([data-disabled='true']):not(\n    [data-selected='true']\n  )\n  .jb-tree-item-overlay {\n  background-color: var(--jb-tree-item-active-bg);\n}\n\n.jb-dynamic-tree\n  .jb-tree-item-button:focus-visible:not([data-disabled='true'])\n  .jb-tree-item-overlay {\n  box-shadow:\n    0 0 0 1px var(--jb-tree-focus-ring-offset),\n    0 0 0 3px var(--jb-tree-focus-ring);\n}\n\n.jb-dynamic-tree\n  .jb-tree-item-button[data-selected='true']\n  .jb-tree-item-overlay {\n  background-color: var(--jb-tree-item-selected-bg);\n}\n\n.jb-dynamic-tree .jb-tree-item-disclosure {\n  color: var(--jb-tree-disclosure);\n}\n\n.jb-dynamic-tree\n  .jb-tree-item-button:not([data-selected='true']):not([data-disabled='true'])\n  .jb-tree-item-disclosure:hover {\n  color: var(--jb-tree-disclosure-hover);\n}\n\n.jb-dynamic-tree\n  .jb-tree-item-button[data-disabled='true']\n  .jb-tree-item-disclosure {\n  color: var(--jb-tree-disabled-text);\n}\n\n.jb-dynamic-tree .jb-tree-item-icon {\n  color: var(--jb-tree-icon);\n}\n\n.jb-dynamic-tree\n  .jb-tree-item-button[data-selected='true']\n  .jb-tree-item-disclosure,\n.jb-dynamic-tree .jb-tree-item-button[data-selected='true'] .jb-tree-item-icon {\n  color: currentColor;\n}\n\n.jb-dynamic-tree .jb-tree-item-end {\n  color: var(--jb-tree-muted);\n}",
       },
     ],
     keywords: ['tree', 'navigation', 'jetbrains', 'ui', 'component'],
@@ -806,6 +820,63 @@ export const index: Record<string, any> = {
       return LazyComp;
     })(),
     command: '@jetbrains-ui/tree',
+  },
+  'dynamic-tree': {
+    name: 'dynamic-tree',
+    description:
+      'JetBrains-style async tree component built on top of @headless-tree with a loadData API.',
+    type: 'registry:ui',
+    dependencies: ['@headless-tree/core', '@headless-tree/react'],
+    devDependencies: undefined,
+    registryDependencies: [
+      'utils',
+      '@jetbrains-ui/icons-components',
+      '@jetbrains-ui/scroll-area',
+    ],
+    files: [
+      {
+        path: 'registry/components/tree/dynamic-tree/index.tsx',
+        type: 'registry:ui',
+        target: 'components/jetbrains-ui/tree/dynamic-tree/index.tsx',
+        content:
+          '\'use client\';\n\nimport * as React from \'react\';\n\nimport {\n  asyncDataLoaderFeature,\n  hotkeysCoreFeature,\n  selectionFeature,\n  type ItemInstance,\n  type TreeDataLoader,\n} from \'@headless-tree/core\';\nimport { useTree } from \'@headless-tree/react\';\nimport {\n  ChevronDownIcon,\n  ChevronRightIcon,\n  LoaderIcon,\n} from \'@/components/jetbrains-ui/icons\';\nimport { ScrollArea, ScrollViewport } from \'@/components/jetbrains-ui/scroll-area\';\nimport \'./index.css\';\n\nconst DYNAMIC_TREE_ROOT_ITEM_ID = \'__jb-dynamic-tree-root__\';\n\nexport interface DynamicTreeItemData {\n  label: string;\n  icon?: React.ReactNode;\n  endContent?: React.ReactNode;\n  disabled?: boolean;\n  isFolder?: boolean;\n}\n\nexport interface DynamicTreeLoadedItem<\n  TItem extends DynamicTreeItemData = DynamicTreeItemData,\n> {\n  id: string;\n  data: TItem;\n}\n\nexport type DynamicTreeLoadData<\n  TItem extends DynamicTreeItemData = DynamicTreeItemData,\n> = (\n  itemId: string | null,\n) => DynamicTreeLoadedItem<TItem>[] | Promise<DynamicTreeLoadedItem<TItem>[]>;\n\nexport interface DynamicTreeProps<\n  TItem extends DynamicTreeItemData = DynamicTreeItemData,\n> extends Omit<\n    React.ComponentPropsWithoutRef<\'div\'>,\n    \'children\' | \'className\' | \'style\'\n  > {\n  loadData: DynamicTreeLoadData<TItem>;\n  treeLabel?: string;\n  indent?: number;\n  width?: React.CSSProperties[\'width\'];\n  height?: React.CSSProperties[\'height\'];\n  loadingLabel?: string;\n}\n\ntype ItemRowProps<TItem extends DynamicTreeItemData> = {\n  indent: number;\n  item: ItemInstance<TItem>;\n  loadingLabel: string;\n};\n\ntype DynamicTreeItemButtonStyle = React.CSSProperties & {\n  \'--jb-tree-item-padding-left\': string;\n};\n\nfunction DynamicTreeItemRow<TItem extends DynamicTreeItemData>({\n  indent,\n  item,\n  loadingLabel,\n}: ItemRowProps<TItem>) {\n  const {\n    onClick: ignoredOnClick,\n    onDoubleClick: ignoredOnDoubleClick,\n    onKeyDown: resolvedOnKeyDown,\n    style: resolvedStyle,\n    tabIndex: resolvedTabIndex,\n    ...resolvedProps\n  } = item.getProps() as React.ComponentPropsWithoutRef<\'div\'>;\n  void ignoredOnClick;\n  void ignoredOnDoubleClick;\n\n  return (\n    <div\n      data-slot="jb-tree-item"\n      data-value={item.getId()}\n      className="jb-tree-item m-0 p-0"\n    >\n      <div\n        {...resolvedProps}\n        data-slot="jb-tree-item-button"\n        data-value={item.getId()}\n        data-has-children={item.isFolder() ? \'true\' : undefined}\n        data-selected={item.isSelected() ? \'true\' : undefined}\n        data-disabled={\n          (item.getItemData() as TItem | null)?.disabled ? \'true\' : undefined\n        }\n        data-expanded={\n          item.isFolder() && item.isExpanded() ? \'true\' : undefined\n        }\n        data-loading={item.isLoading() ? \'true\' : undefined}\n        aria-selected={item.isSelected()}\n        aria-disabled={\n          (item.getItemData() as TItem | null)?.disabled ? true : undefined\n        }\n        aria-busy={item.isLoading() ? true : undefined}\n        tabIndex={\n          (item.getItemData() as TItem | null)?.disabled\n            ? -1\n            : (resolvedTabIndex ?? -1)\n        }\n        style={\n          {\n            ...resolvedStyle,\n            \'--jb-tree-item-padding-left\': `${\n              16 +\n              item.getItemMeta().level * indent +\n              Math.max(0, item.getItemMeta().level - 1) * 2\n            }px`,\n          } as DynamicTreeItemButtonStyle\n        }\n        onClick={() => {\n          if ((item.getItemData() as TItem | null)?.disabled) {\n            return;\n          }\n\n          item.setFocused();\n          item.select();\n        }}\n        onDoubleClick={\n          item.isFolder() && !(item.getItemData() as TItem | null)?.disabled\n            ? () => {\n                item.setFocused();\n\n                if (item.isExpanded()) {\n                  item.collapse();\n                } else {\n                  item.expand();\n                }\n              }\n            : undefined\n        }\n        onKeyDown={(event) => {\n          if ((item.getItemData() as TItem | null)?.disabled) {\n            resolvedOnKeyDown?.(event);\n            return;\n          }\n\n          if (event.key === \'Enter\' || event.key === \' \') {\n            event.preventDefault();\n            item.setFocused();\n            item.select();\n            return;\n          }\n\n          resolvedOnKeyDown?.(event);\n        }}\n        className="jb-tree-item-button relative box-border flex h-6 w-full min-w-0 cursor-default items-center gap-0.5 pr-4 pl-[var(--jb-tree-item-padding-left)] select-none outline-none transition-[color] duration-150 ease-in-out data-[disabled=true]:cursor-not-allowed"\n      >\n        <span\n          aria-hidden="true"\n          data-slot="jb-tree-item-overlay"\n          className="jb-tree-item-overlay pointer-events-none absolute inset-y-0 left-3 right-3 rounded-[4px] transition-[background-color,box-shadow] duration-150 ease-in-out"\n        />\n\n        {item.isFolder() ? (\n          <button\n            data-slot="jb-tree-item-disclosure"\n            data-value={item.getId()}\n            data-expanded={item.isExpanded() ? \'true\' : undefined}\n            data-loading={item.isLoading() ? \'true\' : undefined}\n            type="button"\n            className="jb-tree-item-disclosure relative z-10 inline-flex size-4 shrink-0 items-center justify-center border-0 bg-transparent p-0 outline-none disabled:cursor-not-allowed"\n            aria-label={\n              item.isLoading()\n                ? `Loading ${(item.getItemData() as TItem | null)?.label ?? loadingLabel}`\n                : item.isExpanded()\n                  ? `Collapse ${(item.getItemData() as TItem | null)?.label ?? loadingLabel}`\n                  : `Expand ${(item.getItemData() as TItem | null)?.label ?? loadingLabel}`\n            }\n            onClick={(event) => {\n              event.stopPropagation();\n\n              if ((item.getItemData() as TItem | null)?.disabled) {\n                return;\n              }\n\n              item.setFocused();\n\n              if (item.isExpanded()) {\n                item.collapse();\n              } else {\n                item.expand();\n              }\n            }}\n            disabled={Boolean((item.getItemData() as TItem | null)?.disabled)}\n          >\n            {item.isLoading() ? (\n              <LoaderIcon\n                className="size-4 [animation-duration:900ms]"\n                size={16}\n              />\n            ) : item.isExpanded() ? (\n              <ChevronDownIcon className="size-4 scale-[0.78]" size={16} />\n            ) : (\n              <ChevronRightIcon className="size-4 scale-[0.78]" size={16} />\n            )}\n          </button>\n        ) : (\n          <span\n            data-slot="jb-tree-item-disclosure-placeholder"\n            aria-hidden="true"\n            className="jb-tree-item-disclosure-placeholder relative z-10 inline-flex size-4 shrink-0 items-center justify-center"\n          />\n        )}\n\n        <span\n          data-slot="jb-tree-item-content"\n          className="jb-tree-item-content relative z-10 inline-flex h-5 min-w-0 flex-1 items-center gap-1.5"\n        >\n          {(item.getItemData() as TItem | null)?.icon ? (\n            <span\n              data-slot="jb-tree-item-icon"\n              className="jb-tree-item-icon inline-flex size-4 shrink-0 items-center justify-center [&>svg]:size-4"\n            >\n              {(item.getItemData() as TItem | null)?.icon}\n            </span>\n          ) : null}\n\n          <span\n            data-slot="jb-tree-item-label"\n            className="jb-tree-item-label min-w-0 overflow-hidden text-ellipsis whitespace-nowrap"\n          >\n            {(item.getItemData() as TItem | null)?.label ?? loadingLabel}\n          </span>\n\n          {(item.getItemData() as TItem | null)?.endContent ? (\n            <span\n              data-slot="jb-tree-item-end"\n              className="jb-tree-item-end ml-auto pr-0.5 text-[13px] leading-4 font-medium"\n            >\n              {(item.getItemData() as TItem | null)?.endContent}\n            </span>\n          ) : null}\n        </span>\n      </div>\n    </div>\n  );\n}\n\nfunction DynamicTree<TItem extends DynamicTreeItemData>({\n  loadData,\n  treeLabel = \'Tree\',\n  indent = 16,\n  width,\n  height,\n  loadingLabel = \'Loading...\',\n  ...props\n}: DynamicTreeProps<TItem>) {\n  const loadedItemsRef = React.useRef<Record<string, TItem>>({});\n\n  const dataLoader = React.useMemo<TreeDataLoader<TItem>>(\n    () => ({\n      getItem: async (itemId: string) => {\n        if (itemId === DYNAMIC_TREE_ROOT_ITEM_ID) {\n          return {\n            label: treeLabel,\n            isFolder: true,\n          } as TItem;\n        }\n\n        return (\n          loadedItemsRef.current[itemId] ??\n          ({\n            label: itemId,\n          } as TItem)\n        );\n      },\n      getChildrenWithData: async (itemId: string) =>\n        loadData(itemId === DYNAMIC_TREE_ROOT_ITEM_ID ? null : itemId),\n    }),\n    [loadData, treeLabel],\n  );\n\n  const tree = useTree<TItem>({\n    rootItemId: DYNAMIC_TREE_ROOT_ITEM_ID,\n    dataLoader,\n    isItemFolder: (item) => Boolean(item.getItemData()?.isFolder),\n    getItemName: (item) => item.getItemData()?.label ?? item.getId(),\n    onLoadedItem: (itemId, item) => {\n      loadedItemsRef.current[itemId] = item;\n    },\n    createLoadingItemData: () =>\n      ({\n        label: loadingLabel,\n      }) as TItem,\n    features: [asyncDataLoaderFeature, selectionFeature, hotkeysCoreFeature],\n  });\n\n  const containerProps = tree.getContainerProps(\n    treeLabel,\n  ) as React.ComponentPropsWithoutRef<\'div\'>;\n  const items = tree.getItems();\n\n  return (\n    <ScrollArea\n      data-slot="jb-tree-scroll-area"\n      className="max-w-full"\n      style={{ width, height }}\n    >\n      <ScrollViewport data-slot="jb-tree-scroll-viewport" className="size-full">\n        <div\n          {...containerProps}\n          data-slot="jb-tree"\n          className="jb-tree jb-dynamic-tree block w-full min-w-[244px] box-border py-2 text-[13px] leading-4 font-medium [font-family:var(--jb-font-sans),sans-serif]"\n          {...props}\n        >\n          {items.map((item) => (\n            <DynamicTreeItemRow\n              key={item.getKey()}\n              item={item}\n              indent={indent}\n              loadingLabel={loadingLabel}\n            />\n          ))}\n        </div>\n      </ScrollViewport>\n    </ScrollArea>\n  );\n}\n\nexport { DynamicTree };\n\nexport default DynamicTree;',
+      },
+      {
+        path: 'registry/components/tree/dynamic-tree/index.css',
+        type: 'registry:style',
+        target: 'components/jetbrains-ui/tree/dynamic-tree/index.css',
+        content:
+          ":root {\n  --jb-tree-bg: transparent;\n  --jb-tree-text: var(--jb-gray-1);\n  --jb-tree-muted: var(--jb-gray-7);\n  --jb-tree-icon: var(--jb-gray-6);\n  --jb-tree-disclosure: var(--jb-gray-7);\n  --jb-tree-disclosure-hover: var(--jb-gray-5);\n  --jb-tree-item-hover-bg: var(--jb-gray-12);\n  --jb-tree-item-active-bg: var(--jb-gray-11);\n  --jb-tree-item-selected-bg: var(--jb-blue-11);\n  --jb-tree-item-selected-text: var(--jb-gray-1);\n  --jb-tree-focus-ring: var(--jb-blue-4);\n  --jb-tree-focus-ring-offset: var(--jb-gray-14);\n  --jb-tree-disabled-text: var(--jb-gray-8);\n}\n\n.dark {\n  --jb-tree-bg: transparent;\n  --jb-tree-text: var(--jb-gray-12);\n  --jb-tree-muted: var(--jb-gray-7);\n  --jb-tree-icon: var(--jb-gray-10);\n  --jb-tree-disclosure: var(--jb-gray-10);\n  --jb-tree-disclosure-hover: var(--jb-gray-12);\n  --jb-tree-item-hover-bg: var(--jb-gray-3);\n  --jb-tree-item-active-bg: var(--jb-gray-4);\n  --jb-tree-item-selected-bg: var(--jb-blue-2);\n  --jb-tree-item-selected-text: var(--jb-gray-12);\n  --jb-tree-focus-ring: var(--jb-blue-6);\n  --jb-tree-focus-ring-offset: var(--jb-gray-2);\n  --jb-tree-disabled-text: var(--jb-gray-7);\n}\n\n.jb-dynamic-tree {\n  background-color: var(--jb-tree-bg);\n  color: var(--jb-tree-text);\n}\n\n.jb-dynamic-tree .jb-tree-item-button {\n  color: var(--jb-tree-text);\n}\n\n.jb-dynamic-tree .jb-tree-item-button[data-disabled='true'] {\n  color: var(--jb-tree-disabled-text);\n}\n\n.jb-dynamic-tree .jb-tree-item-button[data-selected='true'] {\n  color: var(--jb-tree-item-selected-text);\n}\n\n.jb-dynamic-tree .jb-tree-item-overlay {\n  background-color: transparent;\n}\n\n.jb-dynamic-tree\n  .jb-tree-item-button:hover:not([data-disabled='true']):not(\n    [data-selected='true']\n  )\n  .jb-tree-item-overlay {\n  background-color: var(--jb-tree-item-hover-bg);\n}\n\n.jb-dynamic-tree\n  .jb-tree-item-button:active:not([data-disabled='true']):not(\n    [data-selected='true']\n  )\n  .jb-tree-item-overlay {\n  background-color: var(--jb-tree-item-active-bg);\n}\n\n.jb-dynamic-tree\n  .jb-tree-item-button:focus-visible:not([data-disabled='true'])\n  .jb-tree-item-overlay {\n  box-shadow:\n    0 0 0 1px var(--jb-tree-focus-ring-offset),\n    0 0 0 3px var(--jb-tree-focus-ring);\n}\n\n.jb-dynamic-tree\n  .jb-tree-item-button[data-selected='true']\n  .jb-tree-item-overlay {\n  background-color: var(--jb-tree-item-selected-bg);\n}\n\n.jb-dynamic-tree .jb-tree-item-disclosure {\n  color: var(--jb-tree-disclosure);\n}\n\n.jb-dynamic-tree\n  .jb-tree-item-button:not([data-selected='true']):not([data-disabled='true'])\n  .jb-tree-item-disclosure:hover {\n  color: var(--jb-tree-disclosure-hover);\n}\n\n.jb-dynamic-tree\n  .jb-tree-item-button[data-disabled='true']\n  .jb-tree-item-disclosure {\n  color: var(--jb-tree-disabled-text);\n}\n\n.jb-dynamic-tree .jb-tree-item-icon {\n  color: var(--jb-tree-icon);\n}\n\n.jb-dynamic-tree\n  .jb-tree-item-button[data-selected='true']\n  .jb-tree-item-disclosure,\n.jb-dynamic-tree .jb-tree-item-button[data-selected='true'] .jb-tree-item-icon {\n  color: currentColor;\n}\n\n.jb-dynamic-tree .jb-tree-item-end {\n  color: var(--jb-tree-muted);\n}",
+      },
+    ],
+    keywords: [
+      'dynamic-tree',
+      'tree',
+      'navigation',
+      'jetbrains',
+      'ui',
+      'component',
+    ],
+    component: (function () {
+      const LazyComp = React.lazy(async () => {
+        const mod = await import(
+          '@/registry/components/tree/dynamic-tree/index.tsx'
+        );
+        const exportName =
+          Object.keys(mod).find(
+            (key) =>
+              typeof mod[key] === 'function' || typeof mod[key] === 'object',
+          ) || 'dynamic-tree';
+        const Comp = mod.default || mod[exportName];
+        if (mod.animations) {
+          (LazyComp as any).animations = mod.animations;
+        }
+        return { default: Comp };
+      });
+      LazyComp.demoProps = {};
+      return LazyComp;
+    })(),
+    command: '@jetbrains-ui/dynamic-tree',
   },
   'demo-alert': {
     name: 'demo-alert',
@@ -1089,44 +1160,6 @@ export const index: Record<string, any> = {
       return LazyComp;
     })(),
     command: '@jetbrains-ui/demo-dropdown-menu',
-  },
-  'demo-dynamic-tree': {
-    name: 'demo-dynamic-tree',
-    description: 'Demo showing dynamic tree component',
-    type: 'registry:ui',
-    dependencies: [],
-    devDependencies: undefined,
-    registryDependencies: ['@jetbrains-ui/tree'],
-    files: [
-      {
-        path: 'registry/demo/components/dynamic-tree/index.tsx',
-        type: 'registry:ui',
-        target: 'components/jetbrains-ui/demo/dynamic-tree.tsx',
-        content:
-          "'use client';\n\nimport * as React from 'react';\n\nimport {\n  DynamicTree,\n  type DynamicTreeContextMenuItem,\n  type DynamicTreeContextMenuNode,\n  type DynamicTreeItemData,\n} from '@/components/jetbrains-ui/tree';\nimport { Markdown, React as ReactFileType } from '@/components/jetbrains-ui/icons/file-types';\nimport { Folder } from '@/components/jetbrains-ui/icons/nodes';\n\nconst dynamicItems: Record<string, DynamicTreeItemData> = {\n  'dynamic-root': {\n    label: 'jetbrains-ui',\n    isFolder: true,\n    icon: <Folder />,\n    endContent: '7',\n  },\n  'dynamic-src': {\n    label: 'src',\n    isFolder: true,\n    icon: <Folder />,\n    endContent: '4',\n  },\n  'dynamic-components': {\n    label: 'components',\n    isFolder: true,\n    icon: <Folder />,\n    endContent: '2',\n  },\n  'dynamic-tree-file': {\n    label: 'tree.tsx',\n    icon: <ReactFileType />,\n  },\n  'dynamic-dynamic-tree-file': {\n    label: 'dynamic-tree.tsx',\n    icon: <ReactFileType />,\n  },\n  'dynamic-lib': {\n    label: 'lib',\n    isFolder: true,\n    icon: <Folder />,\n    endContent: '1',\n  },\n  'dynamic-utils-file': {\n    label: 'utils.ts',\n    icon: <ReactFileType />,\n  },\n  'dynamic-readme': {\n    label: 'README.md',\n    icon: <Markdown />,\n  },\n  'dynamic-package': {\n    label: 'package.json',\n    icon: <ReactFileType />,\n  },\n};\n\nconst dynamicChildren: Record<string, string[]> = {\n  'dynamic-root': ['dynamic-src', 'dynamic-readme', 'dynamic-package'],\n  'dynamic-src': ['dynamic-components', 'dynamic-lib'],\n  'dynamic-components': ['dynamic-tree-file', 'dynamic-dynamic-tree-file'],\n  'dynamic-lib': ['dynamic-utils-file'],\n};\n\nfunction wait(ms: number) {\n  return new Promise<void>((resolve) => {\n    setTimeout(resolve, ms);\n  });\n}\n\nexport default function DynamicTreeDemo() {\n  const [selectedId, setSelectedId] = React.useState<string | null>(null);\n  const [lastAction, setLastAction] = React.useState(\n    'Right click any node to open a shared context menu.',\n  );\n\n  const dataLoader = React.useMemo(\n    () => ({\n      getItem: async (itemId: string) => {\n        await wait(120);\n\n        return (\n          dynamicItems[itemId] ?? {\n            label: itemId,\n          }\n        );\n      },\n      getChildren: async (itemId: string) => {\n        await wait(120);\n\n        return dynamicChildren[itemId] ?? [];\n      },\n    }),\n    [],\n  );\n\n  const contextMenu = React.useCallback(\n    ({\n      id,\n      isFolder,\n      disabled,\n      label,\n    }: DynamicTreeContextMenuNode): DynamicTreeContextMenuItem[] => {\n      const items: DynamicTreeContextMenuItem[] = [\n        {\n          key: 'rename',\n          label: 'Rename',\n          iconName: 'edit',\n          disabled,\n          onSelect: () => setLastAction(`Rename ${label}`),\n          separator: true,\n        },\n      ];\n\n      if (isFolder) {\n        items.push(\n          {\n            key: 'new-file',\n            label: 'New File',\n            iconName: 'add',\n            onSelect: () => setLastAction(`Create file in ${label}`),\n          },\n          {\n            key: 'new-folder',\n            label: 'New Folder',\n            iconName: 'add',\n            onSelect: () => setLastAction(`Create folder in ${label}`),\n            separator: true,\n          },\n        );\n      } else {\n        items.push({\n          key: 'copy-path',\n          label: 'Copy Path',\n          iconName: 'copy',\n          onSelect: () => setLastAction(`Copy path for ${label}`),\n          separator: true,\n        });\n      }\n\n      items.push(\n        {\n          key: 'mark-as',\n          label: 'Mark As',\n          iconName: 'settings',\n          children: [\n            {\n              key: 'favorite',\n              text: 'Favorite',\n              iconName: 'check',\n              onSelect: () => setLastAction(`Mark ${label} as favorite`),\n            },\n            {\n              key: 'excluded',\n              text: 'Excluded',\n              iconName: 'filter',\n              onSelect: () => setLastAction(`Exclude ${label} from search`),\n            },\n          ],\n        },\n        {\n          key: 'refresh',\n          label: 'Reload Children',\n          iconName: 'refresh',\n          disabled: !isFolder,\n          onSelect: () => setLastAction(`Reload ${label}`),\n        },\n        {\n          key: 'delete',\n          label: 'Delete',\n          iconName: 'trash',\n          disabled: id === 'dynamic-root',\n          variant: 'destructive',\n          onSelect: () => setLastAction(`Delete ${label}`),\n        },\n      );\n\n      return items;\n    },\n    [],\n  );\n\n  return (\n    <div className=\"w-[280px] space-y-2\">\n      <DynamicTree\n        className=\"w-full\"\n        rootItemId=\"dynamic-root\"\n        dataLoader={dataLoader}\n        selectedId={selectedId}\n        onSelectedIdChange={setSelectedId}\n        defaultExpandedIds={['dynamic-src']}\n        contextMenu={contextMenu}\n      />\n\n      <div className=\"rounded-[6px] border border-[var(--jb-gray-11)] bg-[var(--jb-gray-14)] px-3 py-2 text-xs text-[var(--jb-gray-7)] dark:border-[var(--jb-gray-4)] dark:bg-[var(--jb-gray-2)] dark:text-[var(--jb-gray-8)]\">\n        {lastAction}\n      </div>\n    </div>\n  );\n}",
-      },
-    ],
-    keywords: [],
-    component: (function () {
-      const LazyComp = React.lazy(async () => {
-        const mod = await import(
-          '@/registry/demo/components/dynamic-tree/index.tsx'
-        );
-        const exportName =
-          Object.keys(mod).find(
-            (key) =>
-              typeof mod[key] === 'function' || typeof mod[key] === 'object',
-          ) || 'demo-dynamic-tree';
-        const Comp = mod.default || mod[exportName];
-        if (mod.animations) {
-          (LazyComp as any).animations = mod.animations;
-        }
-        return { default: Comp };
-      });
-      LazyComp.demoProps = {};
-      return LazyComp;
-    })(),
-    command: '@jetbrains-ui/demo-dynamic-tree',
   },
   'demo-input': {
     name: 'demo-input',
@@ -1547,6 +1580,44 @@ export const index: Record<string, any> = {
     })(),
     command: '@jetbrains-ui/demo-tooltip',
   },
+  'demo-dynamic-tree': {
+    name: 'demo-dynamic-tree',
+    description: 'Demo showing dynamic tree component',
+    type: 'registry:ui',
+    dependencies: [],
+    devDependencies: undefined,
+    registryDependencies: ['@jetbrains-ui/dynamic-tree'],
+    files: [
+      {
+        path: 'registry/demo/components/tree/dynamic-tree/index.tsx',
+        type: 'registry:ui',
+        target: 'components/jetbrains-ui/demo/dynamic-tree.tsx',
+        content:
+          "'use client';\n\nimport * as React from 'react';\n\nimport {\n  DynamicTree,\n  type DynamicTreeItemData,\n  type DynamicTreeLoadedItem,\n  type DynamicTreeLoadData,\n} from '@/components/jetbrains-ui/tree/dynamic-tree';\nimport { Markdown, React as ReactFileType } from '@/components/jetbrains-ui/icons/file-types';\nimport { Folder } from '@/components/jetbrains-ui/icons/nodes';\n\nconst DEMO_ROOT_ID = 'dynamic-root';\n\ntype DemoTreeItem = DynamicTreeItemData & {\n  kind: 'file' | 'folder';\n  path: string;\n};\n\ntype DemoTreeItems = Record<string, Omit<DemoTreeItem, 'endContent'>>;\ntype DemoTreeChildren = Record<string, string[]>;\n\nconst dynamicItems: DemoTreeItems = {\n  [DEMO_ROOT_ID]: {\n    label: 'jetbrains-ui',\n    kind: 'folder',\n    path: '/jetbrains-ui',\n    isFolder: true,\n    icon: <Folder />,\n  },\n  'dynamic-src': {\n    label: 'src',\n    kind: 'folder',\n    path: '/jetbrains-ui/src',\n    isFolder: true,\n    icon: <Folder />,\n  },\n  'dynamic-components': {\n    label: 'components',\n    kind: 'folder',\n    path: '/jetbrains-ui/src/components',\n    isFolder: true,\n    icon: <Folder />,\n  },\n  'dynamic-tree-file': {\n    label: 'tree.tsx',\n    kind: 'file',\n    path: '/jetbrains-ui/src/components/tree.tsx',\n    icon: <ReactFileType />,\n  },\n  'dynamic-dynamic-tree-file': {\n    label: 'dynamic-tree.tsx',\n    kind: 'file',\n    path: '/jetbrains-ui/src/components/dynamic-tree.tsx',\n    icon: <ReactFileType />,\n  },\n  'dynamic-lib': {\n    label: 'lib',\n    kind: 'folder',\n    path: '/jetbrains-ui/src/lib',\n    isFolder: true,\n    icon: <Folder />,\n  },\n  'dynamic-utils-file': {\n    label: 'utils.ts',\n    kind: 'file',\n    path: '/jetbrains-ui/src/lib/utils.ts',\n    icon: <ReactFileType />,\n  },\n  'dynamic-readme': {\n    label: 'README.md',\n    kind: 'file',\n    path: '/jetbrains-ui/README.md',\n    icon: <Markdown />,\n  },\n  'dynamic-package': {\n    label: 'package.json',\n    kind: 'file',\n    path: '/jetbrains-ui/package.json',\n    icon: <ReactFileType />,\n  },\n};\n\nconst dynamicChildren: DemoTreeChildren = {\n  [DEMO_ROOT_ID]: ['dynamic-src', 'dynamic-readme', 'dynamic-package'],\n  'dynamic-src': ['dynamic-components', 'dynamic-lib'],\n  'dynamic-components': ['dynamic-tree-file', 'dynamic-dynamic-tree-file'],\n  'dynamic-lib': ['dynamic-utils-file'],\n};\n\nfunction countDescendants(children: DemoTreeChildren, itemId: string) {\n  const queue = [...(children[itemId] ?? [])];\n  let count = 0;\n\n  while (queue.length > 0) {\n    const currentId = queue.shift();\n\n    if (!currentId) {\n      continue;\n    }\n\n    count += 1;\n    queue.push(...(children[currentId] ?? []));\n  }\n\n  return count;\n}\n\nfunction wait(ms: number) {\n  return new Promise<void>((resolve) => {\n    setTimeout(resolve, ms);\n  });\n}\n\nfunction resolveDemoItemData(\n  itemsById: DemoTreeItems,\n  childrenById: DemoTreeChildren,\n  itemId: string,\n): DemoTreeItem {\n  const item = itemsById[itemId];\n\n  if (!item) {\n    return {\n      label: itemId,\n      kind: 'file',\n      path: itemId,\n    };\n  }\n\n  if (!item.isFolder) {\n    return item;\n  }\n\n  const descendantCount = countDescendants(childrenById, itemId);\n\n  return {\n    ...item,\n    endContent: descendantCount > 0 ? String(descendantCount) : undefined,\n  };\n}\n\nexport default function DynamicTreeDemo() {\n  const loadData = React.useCallback<DynamicTreeLoadData<DemoTreeItem>>(\n    async (itemId) => {\n      await wait(120);\n\n      const parentId = itemId ?? DEMO_ROOT_ID;\n\n      return (dynamicChildren[parentId] ?? []).map(\n        (childId): DynamicTreeLoadedItem<DemoTreeItem> => ({\n          id: childId,\n          data: resolveDemoItemData(dynamicItems, dynamicChildren, childId),\n        }),\n      );\n    },\n    [],\n  );\n\n  return (\n    <div className=\"w-[280px]\">\n      <DynamicTree<DemoTreeItem>\n        width=\"220px\"\n        height=\"120px\"\n        loadData={loadData}\n      />\n    </div>\n  );\n}",
+      },
+    ],
+    keywords: [],
+    component: (function () {
+      const LazyComp = React.lazy(async () => {
+        const mod = await import(
+          '@/registry/demo/components/tree/dynamic-tree/index.tsx'
+        );
+        const exportName =
+          Object.keys(mod).find(
+            (key) =>
+              typeof mod[key] === 'function' || typeof mod[key] === 'object',
+          ) || 'demo-dynamic-tree';
+        const Comp = mod.default || mod[exportName];
+        if (mod.animations) {
+          (LazyComp as any).animations = mod.animations;
+        }
+        return { default: Comp };
+      });
+      LazyComp.demoProps = {};
+      return LazyComp;
+    })(),
+    command: '@jetbrains-ui/demo-dynamic-tree',
+  },
   'demo-tree': {
     name: 'demo-tree',
     description: 'Demo showing tree component',
@@ -1556,17 +1627,19 @@ export const index: Record<string, any> = {
     registryDependencies: ['@jetbrains-ui/tree'],
     files: [
       {
-        path: 'registry/demo/components/tree/index.tsx',
+        path: 'registry/demo/components/tree/tree/index.tsx',
         type: 'registry:ui',
         target: 'components/jetbrains-ui/demo/tree.tsx',
         content:
-          "'use client';\n\nimport * as React from 'react';\n\nimport {\n  Tree,\n  TreeItem,\n  type TreeContextMenuItem,\n  type TreeContextMenuNode,\n} from '@/components/jetbrains-ui/tree';\nimport {\n  Ignored,\n  Markdown,\n  React as ReactFileType,\n} from '@/components/jetbrains-ui/icons/file-types';\nimport { Folder } from '@/components/jetbrains-ui/icons/nodes';\n\ntype TreeDemoState = 'default' | 'selected' | 'disabled' | 'collapsed';\n\ninterface TreeDemoProps {\n  state?: TreeDemoState;\n}\n\nexport default function TreeDemo({ state = 'default' }: TreeDemoProps) {\n  const initialSelectedId = state === 'selected' ? 'tree-file-banner' : null;\n  const [selectedId, setSelectedId] = React.useState<string | null>(\n    initialSelectedId,\n  );\n  const [lastAction, setLastAction] = React.useState(\n    'Right click any node to open a shared context menu.',\n  );\n\n  React.useEffect(() => {\n    setSelectedId(initialSelectedId);\n  }, [initialSelectedId]);\n\n  const lockBannerFile = state === 'disabled';\n  const defaultExpandedIds =\n    state === 'collapsed'\n      ? []\n      : ['tree-root', 'tree-components', 'tree-dialog'];\n\n  const contextMenu = React.useCallback(\n    ({\n      id,\n      label,\n      hasChildren,\n      disabled,\n    }: TreeContextMenuNode): TreeContextMenuItem[] => {\n      const items: TreeContextMenuItem[] = [\n        {\n          key: 'rename',\n          label: 'Rename',\n          iconName: 'edit',\n          disabled,\n          onSelect: () => setLastAction(`Rename ${label}`),\n          separator: true,\n        },\n      ];\n\n      if (hasChildren) {\n        items.push(\n          {\n            key: 'new-file',\n            label: 'New File',\n            iconName: 'add',\n            onSelect: () => setLastAction(`Create file in ${label}`),\n          },\n          {\n            key: 'new-folder',\n            label: 'New Folder',\n            iconName: 'add',\n            onSelect: () => setLastAction(`Create folder in ${label}`),\n            separator: true,\n          },\n        );\n      } else {\n        items.push({\n          key: 'copy-path',\n          label: 'Copy Path',\n          iconName: 'copy',\n          onSelect: () => setLastAction(`Copy path for ${label}`),\n          separator: true,\n        });\n      }\n\n      items.push(\n        {\n          key: 'mark-as',\n          label: 'Mark As',\n          iconName: 'settings',\n          children: [\n            {\n              key: 'favorite',\n              label: 'Favorite',\n              iconName: 'check',\n              onSelect: () => setLastAction(`Mark ${label} as favorite`),\n            },\n            {\n              key: 'excluded',\n              label: 'Excluded',\n              iconName: 'filter',\n              onSelect: () => setLastAction(`Exclude ${label} from search`),\n            },\n          ],\n        },\n        {\n          key: 'refresh',\n          label: 'Reload Children',\n          iconName: 'refresh',\n          disabled: !hasChildren,\n          onSelect: () => setLastAction(`Reload ${label}`),\n        },\n        {\n          key: 'delete',\n          label: 'Delete',\n          iconName: 'trash',\n          disabled: id === 'tree-root',\n          variant: 'destructive',\n          onSelect: () => setLastAction(`Delete ${label}`),\n        },\n      );\n\n      return items;\n    },\n    [],\n  );\n\n  return (\n    <div className=\"w-[280px] space-y-2\">\n      <Tree\n        className=\"w-full\"\n        selectedId={selectedId}\n        onSelectedIdChange={setSelectedId}\n        defaultExpandedIds={defaultExpandedIds}\n        contextMenu={contextMenu}\n      >\n        <TreeItem\n          value=\"tree-root\"\n          label=\"jetbrains-ui\"\n          icon={<Folder />}\n          endContent=\"12\"\n        >\n          <TreeItem value=\"tree-git\" label=\".gitignore\" icon={<Ignored />} />\n\n          <TreeItem\n            value=\"tree-components\"\n            label=\"components\"\n            icon={<Folder />}\n            endContent=\"4\"\n          >\n            <TreeItem\n              value=\"tree-dialog\"\n              label=\"dialog\"\n              icon={<Folder />}\n              endContent=\"2\"\n            >\n              <TreeItem\n                value=\"tree-file-banner\"\n                label=\"banner.tsx\"\n                icon={<ReactFileType />}\n                disabled={lockBannerFile}\n              />\n              <TreeItem\n                value=\"tree-file-rd-dialog\"\n                label=\"rd-dialog.tsx\"\n                icon={<ReactFileType />}\n              />\n            </TreeItem>\n\n            <TreeItem\n              value=\"tree-tree\"\n              label=\"tree.tsx\"\n              icon={<ReactFileType />}\n            />\n            <TreeItem\n              value=\"tree-toggle\"\n              label=\"toggle.tsx\"\n              icon={<ReactFileType />}\n            />\n          </TreeItem>\n\n          <TreeItem value=\"tree-readme\" label=\"README.md\" icon={<Markdown />} />\n        </TreeItem>\n      </Tree>\n\n      <div className=\"rounded-[6px] border border-[var(--jb-gray-11)] bg-[var(--jb-gray-14)] px-3 py-2 text-xs text-[var(--jb-gray-7)] dark:border-[var(--jb-gray-4)] dark:bg-[var(--jb-gray-2)] dark:text-[var(--jb-gray-8)]\">\n        {lastAction}\n      </div>\n    </div>\n  );\n}",
+          '\'use client\';\n\nimport * as React from \'react\';\n\nimport { Tree, TreeItem } from \'@/components/jetbrains-ui/tree\';\nimport {\n  Ignored,\n  Markdown,\n  React as ReactFileType,\n} from \'@/components/jetbrains-ui/icons/file-types\';\nimport { Folder } from \'@/components/jetbrains-ui/icons/nodes\';\n\ntype TreeDemoState = \'default\' | \'selected\' | \'disabled\' | \'collapsed\';\n\ninterface TreeDemoProps {\n  state?: TreeDemoState;\n}\n\nexport default function TreeDemo({ state = \'default\' }: TreeDemoProps) {\n  const initialSelectedId = state === \'selected\' ? \'tree-file-banner\' : null;\n  const [selectedId, setSelectedId] = React.useState<string | null>(\n    initialSelectedId,\n  );\n\n  React.useEffect(() => {\n    setSelectedId(initialSelectedId);\n  }, [initialSelectedId]);\n\n  const lockBannerFile = state === \'disabled\';\n  const defaultExpandedIds =\n    state === \'collapsed\'\n      ? []\n      : [\'tree-root\', \'tree-components\', \'tree-dialog\'];\n\n  return (\n    <div className="w-[280px] space-y-2">\n      <Tree\n        className="w-full"\n        width="220px"\n        height="156px"\n        selectedId={selectedId}\n        onSelectedIdChange={setSelectedId}\n        defaultExpandedIds={defaultExpandedIds}\n      >\n        <TreeItem\n          value="tree-root"\n          label="jetbrains-ui"\n          icon={<Folder />}\n          endContent="12"\n        >\n          <TreeItem value="tree-git" label=".gitignore" icon={<Ignored />} />\n\n          <TreeItem\n            value="tree-components"\n            label="components"\n            icon={<Folder />}\n            endContent="4"\n          >\n            <TreeItem\n              value="tree-dialog"\n              label="dialog"\n              icon={<Folder />}\n              endContent="2"\n            >\n              <TreeItem\n                value="tree-file-banner"\n                label="banner.tsx"\n                icon={<ReactFileType />}\n                disabled={lockBannerFile}\n              />\n              <TreeItem\n                value="tree-file-rd-dialog"\n                label="rd-dialog.tsx"\n                icon={<ReactFileType />}\n              />\n            </TreeItem>\n\n            <TreeItem\n              value="tree-tree"\n              label="tree.tsx"\n              icon={<ReactFileType />}\n            />\n            <TreeItem\n              value="tree-toggle"\n              label="toggle.tsx"\n              icon={<ReactFileType />}\n            />\n          </TreeItem>\n\n          <TreeItem value="tree-readme" label="README.md" icon={<Markdown />} />\n        </TreeItem>\n      </Tree>\n    </div>\n  );\n}',
       },
     ],
     keywords: [],
     component: (function () {
       const LazyComp = React.lazy(async () => {
-        const mod = await import('@/registry/demo/components/tree/index.tsx');
+        const mod = await import(
+          '@/registry/demo/components/tree/tree/index.tsx'
+        );
         const exportName =
           Object.keys(mod).find(
             (key) =>
