--- apps/www/content/docs/components/tree.mdx
@@ -1,96 +1,0 @@
----
-title: Tree
-description: Tree component styled with the JetBrains UI design language.
-author:
-  name: minesunny
-  url: https://github.com/minesunny
-releaseDate: 2026-03-21
----
-
-<ComponentPreview name="demo-tree" />
-
-## Overview
-
-A JetBrains-style tree view for hierarchical navigation.
-
-## Installation
-
-<ComponentInstallation name="tree" />
-
-## Usage
-
-```tsx
-import { Tree, TreeItem } from '@/components/jetbrains-ui/tree';
-
-function Example() {
-  return (
-    <Tree defaultExpandedIds={['root']}>
-      <TreeItem value="root" label="project">
-        <TreeItem value="src" label="src">
-          <TreeItem value="index" label="index.tsx" />
-        </TreeItem>
-      </TreeItem>
-    </Tree>
-  );
-}
-```
-
-### DynamicTree
-
-<ComponentPreview name="demo-dynamic-tree" />
-
-```tsx
-import { DynamicTree } from '@/components/jetbrains-ui/tree';
-
-const dataLoader = {
-  async getItem(itemId: string) {
-    return {
-      label: itemId,
-      isFolder: itemId.startsWith('folder-'),
-    };
-  },
-  async getChildren(itemId: string) {
-    if (itemId === 'root') return ['folder-src', 'readme-md'];
-    if (itemId === 'folder-src') return ['index-tsx'];
-    return [];
-  },
-};
-
-function DynamicExample() {
-  return (
-    <DynamicTree
-      rootItemId="root"
-      dataLoader={dataLoader}
-      defaultExpandedIds={['folder-src']}
-      contextMenu={(node) => [
-        {
-          key: 'rename',
-          label: 'Rename',
-          iconName: 'edit',
-          onSelect: () => console.log('rename', node.id),
-        },
-        {
-          key: 'more',
-          label: 'More',
-          children: [
-            {
-              key: 'copy-path',
-              label: 'Copy Path',
-              iconName: 'copy',
-            },
-          ],
-        },
-      ]}
-    />
-  );
-}
-```
-
-## API Highlights
-
-- `Tree`: root container, supports controlled and uncontrolled selected and expanded state.
-- `TreeItem`: static tree node with optional children, icon and end content.
-- `Tree.contextMenu`: shared right-click callback for static tree nodes.
-- `DynamicTree`: built on `@headless-tree`, supports async `dataLoader` and dynamic node rendering.
-- `DynamicTree.contextMenu`: shared right-click callback that receives the current dynamic node.
-- Built-in selection, collapse toggle and keyboard support with `Enter`, `Space`, `ArrowLeft` and `ArrowRight`.
