+++ apps/www/registry/components/icons/index.ts
@@ -11,6 +11,7 @@ export {
   TrashIcon,
   EditIcon,
   RefreshIcon,
+  LoaderIcon,
   FilterIcon,
 } from './icons';
