+++ apps/www/registry/components/tree/dynamic-tree/index.css
@@ -1,0 +1,113 @@
+:root {
+  --jb-tree-bg: transparent;
+  --jb-tree-text: var(--jb-gray-1);
+  --jb-tree-muted: var(--jb-gray-7);
+  --jb-tree-icon: var(--jb-gray-6);
+  --jb-tree-disclosure: var(--jb-gray-7);
+  --jb-tree-disclosure-hover: var(--jb-gray-5);
+  --jb-tree-item-hover-bg: var(--jb-gray-12);
+  --jb-tree-item-active-bg: var(--jb-gray-11);
+  --jb-tree-item-selected-bg: var(--jb-blue-11);
+  --jb-tree-item-selected-text: var(--jb-gray-1);
+  --jb-tree-focus-ring: var(--jb-blue-4);
+  --jb-tree-focus-ring-offset: var(--jb-gray-14);
+  --jb-tree-disabled-text: var(--jb-gray-8);
+}
+
+.dark {
+  --jb-tree-bg: transparent;
+  --jb-tree-text: var(--jb-gray-12);
+  --jb-tree-muted: var(--jb-gray-7);
+  --jb-tree-icon: var(--jb-gray-10);
+  --jb-tree-disclosure: var(--jb-gray-10);
+  --jb-tree-disclosure-hover: var(--jb-gray-12);
+  --jb-tree-item-hover-bg: var(--jb-gray-3);
+  --jb-tree-item-active-bg: var(--jb-gray-4);
+  --jb-tree-item-selected-bg: var(--jb-blue-2);
+  --jb-tree-item-selected-text: var(--jb-gray-12);
+  --jb-tree-focus-ring: var(--jb-blue-6);
+  --jb-tree-focus-ring-offset: var(--jb-gray-2);
+  --jb-tree-disabled-text: var(--jb-gray-7);
+}
+
+.jb-dynamic-tree {
+  background-color: var(--jb-tree-bg);
+  color: var(--jb-tree-text);
+}
+
+.jb-dynamic-tree .jb-tree-item-button {
+  color: var(--jb-tree-text);
+}
+
+.jb-dynamic-tree .jb-tree-item-button[data-disabled='true'] {
+  color: var(--jb-tree-disabled-text);
+}
+
+.jb-dynamic-tree .jb-tree-item-button[data-selected='true'] {
+  color: var(--jb-tree-item-selected-text);
+}
+
+.jb-dynamic-tree .jb-tree-item-overlay {
+  background-color: transparent;
+}
+
+.jb-dynamic-tree
+  .jb-tree-item-button:hover:not([data-disabled='true']):not(
+    [data-selected='true']
+  )
+  .jb-tree-item-overlay {
+  background-color: var(--jb-tree-item-hover-bg);
+}
+
+.jb-dynamic-tree
+  .jb-tree-item-button:active:not([data-disabled='true']):not(
+    [data-selected='true']
+  )
+  .jb-tree-item-overlay {
+  background-color: var(--jb-tree-item-active-bg);
+}
+
+.jb-dynamic-tree
+  .jb-tree-item-button:focus-visible:not([data-disabled='true'])
+  .jb-tree-item-overlay {
+  box-shadow:
+    0 0 0 1px var(--jb-tree-focus-ring-offset),
+    0 0 0 3px var(--jb-tree-focus-ring);
+}
+
+.jb-dynamic-tree
+  .jb-tree-item-button[data-selected='true']
+  .jb-tree-item-overlay {
+  background-color: var(--jb-tree-item-selected-bg);
+}
+
+.jb-dynamic-tree .jb-tree-item-disclosure {
+  color: var(--jb-tree-disclosure);
+}
+
+.jb-dynamic-tree
+  .jb-tree-item-button:not([data-selected='true']):not([data-disabled='true'])
+  .jb-tree-item-disclosure:hover {
+  color: var(--jb-tree-disclosure-hover);
+}
+
+.jb-dynamic-tree
+  .jb-tree-item-button[data-disabled='true']
+  .jb-tree-item-disclosure {
+  color: var(--jb-tree-disabled-text);
+}
+
+.jb-dynamic-tree .jb-tree-item-icon {
+  color: var(--jb-tree-icon);
+}
+
+.jb-dynamic-tree
+  .jb-tree-item-button[data-selected='true']
+  .jb-tree-item-disclosure,
+.jb-dynamic-tree .jb-tree-item-button[data-selected='true'] .jb-tree-item-icon {
+  color: currentColor;
+}
+
+.jb-dynamic-tree .jb-tree-item-end {
+  color: var(--jb-tree-muted);
+}
