+++ apps/www/registry/components/tree/tree/index.css
@@ -30,13 +30,13 @@
   --jb-tree-disabled-text: var(--jb-gray-7);
 }
 
-.jb-tree {
+.jb-static-tree {
   background: var(--jb-tree-bg);
   color: var(--jb-tree-text);
   font-family: var(--jb-font-sans), sans-serif;
 }
 
-.jb-tree-group-motion {
+.jb-static-tree .jb-tree-group-motion {
   display: grid;
   grid-template-rows: 0fr;
   min-height: 0;
@@ -49,77 +49,77 @@
     transform 170ms var(--jb-ease-out);
 }
 
-.jb-tree-group-motion[data-expanded='true'] {
+.jb-static-tree .jb-tree-group-motion[data-expanded='true'] {
   grid-template-rows: 1fr;
   opacity: 1;
   transform: translateY(0);
   pointer-events: auto;
 }
 
-.jb-tree-group-motion .jb-tree-group {
+.jb-static-tree .jb-tree-group-motion .jb-tree-group {
   min-height: 0;
   overflow: hidden;
 }
 
-.jb-tree-item-button {
+.jb-static-tree .jb-tree-item-button {
   position: relative;
   color: var(--jb-tree-text);
   box-sizing: border-box;
 }
 
-.jb-tree-item-button::before {
-  content: '';
-  position: absolute;
-  inset: 0 12px;
-  border-radius: 4px;
-  background: transparent;
-  transition:
-    background-color var(--jb-transition-fast) ease-in-out,
-    box-shadow var(--jb-transition-fast) ease-in-out;
+.jb-static-tree .jb-tree-item-overlay {
+  background-color: transparent;
 }
 
-.jb-tree-item-button > * {
-  position: relative;
-  z-index: 1;
-}
-
-.jb-tree-item-button:hover:not([data-disabled='true']) {
+.jb-static-tree .jb-tree-item-button:hover:not([data-disabled='true']) {
   background: transparent;
 }
 
-.jb-tree-item-button:hover:not([data-disabled='true'])::before {
+.jb-static-tree
+  .jb-tree-item-button:hover:not([data-disabled='true']):not(
+    [data-selected='true']
+  )
+  .jb-tree-item-overlay {
   background: var(--jb-tree-item-hover-bg);
 }
 
-.jb-tree-item-button:active:not([data-disabled='true']) {
+.jb-static-tree .jb-tree-item-button:active:not([data-disabled='true']) {
   background: transparent;
 }
 
-.jb-tree-item-button:active:not([data-disabled='true'])::before {
+.jb-static-tree
+  .jb-tree-item-button:active:not([data-disabled='true']):not(
+    [data-selected='true']
+  )
+  .jb-tree-item-overlay {
   background: var(--jb-tree-item-active-bg);
 }
 
-.jb-tree-item-button[data-selected='true'] {
+.jb-static-tree .jb-tree-item-button[data-selected='true'] {
   background: transparent;
   color: var(--jb-tree-item-selected-text);
 }
 
-.jb-tree-item-button[data-selected='true']::before {
+.jb-static-tree
+  .jb-tree-item-button[data-selected='true']
+  .jb-tree-item-overlay {
   background: var(--jb-tree-item-selected-bg);
 }
 
-.jb-tree-item-button:focus-visible::before {
+.jb-static-tree
+  .jb-tree-item-button:focus-visible:not([data-disabled='true'])
+  .jb-tree-item-overlay {
   box-shadow:
     0 0 0 1px var(--jb-tree-focus-ring-offset),
     0 0 0 3px var(--jb-tree-focus-ring);
 }
 
-.jb-tree-item-button[data-disabled='true'] {
+.jb-static-tree .jb-tree-item-button[data-disabled='true'] {
   color: var(--jb-tree-disabled-text);
   cursor: not-allowed;
 }
 
-.jb-tree-item-disclosure {
+.jb-static-tree .jb-tree-item-disclosure {
   border: 0;
   background: transparent;
   padding: 0;
@@ -127,35 +127,33 @@
   outline: none;
 }
 
-.jb-tree-item-disclosure:hover:not(:disabled) {
-  color: var(--jb-tree-disclosure-hover);
-}
-
-.jb-tree-item-disclosure:disabled {
+.jb-static-tree .jb-tree-item-disclosure:disabled {
   cursor: not-allowed;
 }
 
-.jb-tree-item-chevron {
-  width: 16px;
-  height: 16px;
-  transform: scale(0.78);
-  transform-origin: center;
+.jb-static-tree
+  .jb-tree-item-button:not([data-selected='true']):not([data-disabled='true'])
+  .jb-tree-item-disclosure:hover {
+  color: var(--jb-tree-disclosure-hover);
 }
 
-.jb-tree-item-icon {
+.jb-static-tree .jb-tree-item-icon {
   color: var(--jb-tree-icon);
 }
 
-.jb-tree-item-button[data-selected='true'] .jb-tree-item-icon,
-.jb-tree-item-button[data-selected='true'] .jb-tree-item-disclosure {
+.jb-static-tree
+  .jb-tree-item-button[data-disabled='true']
+  .jb-tree-item-disclosure {
+  color: var(--jb-tree-disabled-text);
+}
+
+.jb-static-tree .jb-tree-item-button[data-selected='true'] .jb-tree-item-icon,
+.jb-static-tree
+  .jb-tree-item-button[data-selected='true']
+  .jb-tree-item-disclosure {
   color: currentColor;
 }
 
-.jb-tree-item-icon > svg {
-  width: 16px;
-  height: 16px;
-}
-
-.jb-tree-item-end {
+.jb-static-tree .jb-tree-item-end {
   color: var(--jb-tree-muted);
 }
