+++ apps/www/registry/demo/components/tree/dynamic-tree/registry-item.json
@@ -5,10 +5,10 @@
   "title": "Dynamic Tree Demo",
   "description": "Demo showing dynamic tree component",
   "dependencies": [],
-  "registryDependencies": ["@jetbrains-ui/tree"],
+  "registryDependencies": ["@jetbrains-ui/dynamic-tree"],
   "files": [
     {
-      "path": "registry/demo/components/dynamic-tree/index.tsx",
+      "path": "registry/demo/components/tree/dynamic-tree/index.tsx",
       "type": "registry:ui",
       "target": "components/jetbrains-ui/demo/dynamic-tree.tsx"
     }
