--- apps/www/registry/demo/components/tree/index.tsx
@@ -1,192 +1,0 @@
-'use client';
-
-import * as React from 'react';
-
-import {
-  Tree,
-  TreeItem,
-  type TreeContextMenuItem,
-  type TreeContextMenuNode,
-} from '@/registry/components/tree';
-import {
-  Ignored,
-  Markdown,
-  React as ReactFileType,
-} from '@/registry/icons/file-types';
-import { Folder } from '@/registry/icons/nodes';
-
-type TreeDemoState = 'default' | 'selected' | 'disabled' | 'collapsed';
-
-interface TreeDemoProps {
-  state?: TreeDemoState;
-}
-
-export default function TreeDemo({ state = 'default' }: TreeDemoProps) {
-  const initialSelectedId = state === 'selected' ? 'tree-file-banner' : null;
-  const [selectedId, setSelectedId] = React.useState<string | null>(
-    initialSelectedId,
-  );
-  const [lastAction, setLastAction] = React.useState(
-    'Right click any node to open a shared context menu.',
-  );
-
-  React.useEffect(() => {
-    setSelectedId(initialSelectedId);
-  }, [initialSelectedId]);
-
-  const lockBannerFile = state === 'disabled';
-  const defaultExpandedIds =
-    state === 'collapsed'
-      ? []
-      : ['tree-root', 'tree-components', 'tree-dialog'];
-
-  const contextMenu = React.useCallback(
-    ({
-      id,
-      label,
-      hasChildren,
-      disabled,
-    }: TreeContextMenuNode): TreeContextMenuItem[] => {
-      const items: TreeContextMenuItem[] = [
-        {
-          key: 'rename',
-          label: 'Rename',
-          iconName: 'edit',
-          disabled,
-          onSelect: () => setLastAction(`Rename ${label}`),
-          separator: true,
-        },
-      ];
-
-      if (hasChildren) {
-        items.push(
-          {
-            key: 'new-file',
-            label: 'New File',
-            iconName: 'add',
-            onSelect: () => setLastAction(`Create file in ${label}`),
-          },
-          {
-            key: 'new-folder',
-            label: 'New Folder',
-            iconName: 'add',
-            onSelect: () => setLastAction(`Create folder in ${label}`),
-            separator: true,
-          },
-        );
-      } else {
-        items.push({
-          key: 'copy-path',
-          label: 'Copy Path',
-          iconName: 'copy',
-          onSelect: () => setLastAction(`Copy path for ${label}`),
-          separator: true,
-        });
-      }
-
-      items.push(
-        {
-          key: 'mark-as',
-          label: 'Mark As',
-          iconName: 'settings',
-          children: [
-            {
-              key: 'favorite',
-              label: 'Favorite',
-              iconName: 'check',
-              onSelect: () => setLastAction(`Mark ${label} as favorite`),
-            },
-            {
-              key: 'excluded',
-              label: 'Excluded',
-              iconName: 'filter',
-              onSelect: () => setLastAction(`Exclude ${label} from search`),
-            },
-          ],
-        },
-        {
-          key: 'refresh',
-          label: 'Reload Children',
-          iconName: 'refresh',
-          disabled: !hasChildren,
-          onSelect: () => setLastAction(`Reload ${label}`),
-        },
-        {
-          key: 'delete',
-          label: 'Delete',
-          iconName: 'trash',
-          disabled: id === 'tree-root',
-          variant: 'destructive',
-          onSelect: () => setLastAction(`Delete ${label}`),
-        },
-      );
-
-      return items;
-    },
-    [],
-  );
-
-  return (
-    <div className="w-[280px] space-y-2">
-      <Tree
-        className="w-full"
-        selectedId={selectedId}
-        onSelectedIdChange={setSelectedId}
-        defaultExpandedIds={defaultExpandedIds}
-        contextMenu={contextMenu}
-      >
-        <TreeItem
-          value="tree-root"
-          label="jetbrains-ui"
-          icon={<Folder />}
-          endContent="12"
-        >
-          <TreeItem value="tree-git" label=".gitignore" icon={<Ignored />} />
-
-          <TreeItem
-            value="tree-components"
-            label="components"
-            icon={<Folder />}
-            endContent="4"
-          >
-            <TreeItem
-              value="tree-dialog"
-              label="dialog"
-              icon={<Folder />}
-              endContent="2"
-            >
-              <TreeItem
-                value="tree-file-banner"
-                label="banner.tsx"
-                icon={<ReactFileType />}
-                disabled={lockBannerFile}
-              />
-              <TreeItem
-                value="tree-file-rd-dialog"
-                label="rd-dialog.tsx"
-                icon={<ReactFileType />}
-              />
-            </TreeItem>
-
-            <TreeItem
-              value="tree-tree"
-              label="tree.tsx"
-              icon={<ReactFileType />}
-            />
-            <TreeItem
-              value="tree-toggle"
-              label="toggle.tsx"
-              icon={<ReactFileType />}
-            />
-          </TreeItem>
-
-          <TreeItem value="tree-readme" label="README.md" icon={<Markdown />} />
-        </TreeItem>
-      </Tree>
-
-      <div className="rounded-[6px] border border-[var(--jb-gray-11)] bg-[var(--jb-gray-14)] px-3 py-2 text-xs text-[var(--jb-gray-7)] dark:border-[var(--jb-gray-4)] dark:bg-[var(--jb-gray-2)] dark:text-[var(--jb-gray-8)]">
-        {lastAction}
-      </div>
-    </div>
-  );
-}
