+++ apps/www/src/__tests__/setup.ts
@@ -1,2 +1,10 @@
 /// <reference types="vitest/globals" />
 import '@testing-library/jest-dom/vitest';
+
+class ResizeObserverMock implements ResizeObserver {
+  observe() {}
+  unobserve() {}
+  disconnect() {}
+}
+
+vi.stubGlobal('ResizeObserver', ResizeObserverMock);
